% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{add_boot_datasets}
\alias{add_boot_datasets}
\title{add/write bootstrap datasets}
\usage{
add_boot_datasets(
  dboots,
  d,
  folder = "DerivedData/bootstrap_datasets",
  overwrite = FALSE
)
}
\arguments{
\item{dboots}{rsplits dataset with run_id column}

\item{d}{dataset to merge}

\item{folder}{folder to store datasets}

\item{overwrite}{overwrite or not}
}
\description{
preparation step before creating nm models
}
