% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise_covariate.R
\name{add_remove_covs}
\alias{add_remove_covs}
\alias{add_cov}
\alias{remove_cov}
\title{Add/remove a covariate to a NONMEM model}
\usage{
add_cov(
  ctl,
  param,
  cov,
  state = 2,
  continuous = TRUE,
  time_varying,
  additional_state_text,
  id_var = "ID",
  force = FALSE,
  force_TV_var = FALSE,
  init,
  lower,
  upper
)

remove_cov(
  ctl,
  param,
  cov,
  state = 2,
  continuous = TRUE,
  time_varying,
  additional_state_text,
  id_var = "ID"
)
}
\arguments{
\item{ctl}{An nm object or an object coercible to ctl_list.}

\item{param}{Character. Name of parameter.}

\item{cov}{Character. Name of covariate.}

\item{state}{Numeric or character. Number or name of state (see details).}

\item{continuous}{Logical (default = TRUE). is covariate continuous?}

\item{time_varying}{Optional logical. is the covariate time varying?}

\item{additional_state_text}{Optional character. custom state variable to be
passed to \code{param_cov_text}.}

\item{id_var}{Character (default = "ID"). Needed if time_varying is missing.}

\item{force}{Logical (default = FALSE). Force covariate in even if missing
values found.}

\item{force_TV_var}{Logical (default = FALSE). Force covariates only on TV
notation parameters.}

\item{init}{Optional numeric/character vector.  Initial estimate of
additional parameters.}

\item{lower}{Optional numeric/character vector.  lower bound of additional
parameters.}

\item{upper}{Optional numeric/character vector.  Upper bound of additional
parameters.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Follows PsN coding conventions to add covariates into a model.  The advantage
is no need to create a .scm file, just directly modify the ctl file contents.
This function is used by \code{\link[=covariate_step_tibble]{covariate_step_tibble()}} for stepwise
covariate model development.
}
\details{
available \code{state}s:

\describe{

\item{"2" or "linear"}{
PARCOV= ( 1 + THETA(1)*(COV -median))
}

\item{"3" or "hockey-stick"}{
IF(COV.LE.median) PARCOV = ( 1 + THETA(1)*(COV - median))
IF(COV.GT.median) PARCOV = ( 1 + THETA(2)*(COV - median))
}

\item{"4" or "exponential"}{
PARCOV= EXP(THETA(1)*(COV - median))
}

\item{"5" or "power"}{
PARCOV= ((COV/median)**THETA(1))
}

\item{"power1"}{
PARCOV= ((COV/median))
}

\item{"power0.75"}{
PARCOV= ((COV/median)**0.75)
}

\item{"6" or "log-linear"}{
PARCOV= ( 1 + THETA(1)*(LOG(COV) - log(median)))
}

}

\code{remove_cov} only works with covariates added with \code{add_cov}.
}
\examples{
\dontrun{

m1WT <- m1 \%>\% child("m1WT") \%>\%
  add_cov(param = "CL", cov = "WT", state = "power") \%>\%
  run_nm()

## compare results

rr(c(m1, m1WT))
summary_wide(c(m1, m1WT))

}


\dontrun{

m1noWT <- m1 \%>\% child("m1noWT") \%>\%
  remove_cov(param = "CL", cov = "WT") \%>\% 
  run_nm()
  
## compare results

rr(c(m1, m1noWT))
summary_wide(c(m1, m1noWT))
  
}

}
\seealso{
\code{\link[=covariate_step_tibble]{covariate_step_tibble()}}, \code{\link[=test_relations]{test_relations()}}
}
