% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subroutine.R
\name{advan}
\alias{advan}
\alias{trans}
\alias{tol}
\title{Get/set ADVAN/TRANS/TOL in ctl_contents}
\usage{
advan(m, text)

trans(m, text)

tol(m, text)
}
\arguments{
\item{m}{nm object}

\item{text}{optional number/character number to set to}
}
\description{
Will make no other changes to control file
}
