% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_record.R
\name{AIC_nm}
\alias{AIC_nm}
\title{Get BIC}
\usage{
AIC_nm(object, ..., k = 2)
}
\arguments{
\item{object}{nm object or list of nm objects}

\item{...}{additional arguments}

\item{k}{integer. default = 2. AIC penalty}
}
\description{
Get BIC
}
