% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_building.R
\name{append_nonmem_var}
\alias{append_nonmem_var}
\title{Include NONMEM variables in output table}
\usage{
append_nonmem_var(output_table, r, var)
}
\arguments{
\item{output_table}{output from output_table()}

\item{r}{nm object}

\item{var}{character. name of variable to compute (needs to be defined in
$PK/$PRED)}
}
\description{
Experimental function - designed to be used in situation where you don't want
to rerun NONMEM, but need a variable in the dataset that wasn't specified in
$TABLE
}
