% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_manual_edit.R
\name{apply_manual_edit}
\alias{apply_manual_edit}
\title{Apply a manual edit patch}
\usage{
apply_manual_edit(m, patch_name)
}
\arguments{
\item{m}{nm object}

\item{patch_name}{character name of patch}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Best to allow the \code{"manual edit"} RStudio addin to write this function in
your script for you.  After a tracked manual edit is performed, a patch file
is created and saved in \code{Models/patches}.  This function applies the patch to
the object.
}
\details{
Generally best to to apply patches before automatic edits and
changes in directories e.g. via \code{run_in()}
}
