% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_methods.R
\name{as_nm_generic}
\alias{as_nm_generic}
\title{Convert nm object to nm_generic}
\usage{
as_nm_generic(m)
}
\arguments{
\item{m}{An \code{nm_list} object to coerce.}
}
\value{
An \code{nm_generic} object.
}
\description{
Used where methods for \code{nm_list} don't exist or aren't appropriate.
}
\keyword{internal}
