% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_methods.R
\name{as_nm_list}
\alias{as_nm_list}
\title{Coerce object into nm_list}
\usage{
as_nm_list(m)
}
\arguments{
\item{m}{An \code{nm_generic} or \code{list} object to coerce.}
}
\value{
An \code{nm_list} object.
}
\description{
This may be relegated to a back end function soon.  Create an \code{nm_list} from
an \code{nm_generic} or \code{list} or other object coercible into an \code{nm_list}.
}
\keyword{internal}
