% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMproject.R
\name{attach_code_library}
\alias{attach_code_library}
\title{Attach local package library to code_library_paths}
\usage{
attach_code_library()
}
\description{
Attach local package library to code_library_paths
}

