% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-gettersetters.R
\name{based_on}
\alias{based_on}
\title{Use file to set control file contents in nm object}
\usage{
based_on(m, ctl_ob, update_numbering = TRUE, update_dollar_data = TRUE, ...)
}
\arguments{
\item{m}{An nm object.}

\item{ctl_ob}{Path to control file.}

\item{update_numbering}{Logical. Should table numbers and author fields be updated.}

\item{update_dollar_data}{Logical. Should $DATA in control file be updated.}

\item{...}{Additional arguments.}
}
\value{
An nm object with modified \code{ctl_contents} field.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function is mostly called by the \code{new_nm()} and will not generally be
used directly by users. The function populates/modifies the \code{ctl_contents}
field of an nm object, with the contents of an already existing control file.
}
\keyword{internal}
