% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise_covariate.R
\name{bind_covariate_results}
\alias{bind_covariate_results}
\title{Add run results into a covariate tibble}
\usage{
bind_covariate_results(dsc, nm_col = "m", parameters = "new")
}
\arguments{
\item{dsc}{A covariate tibble (see \code{\link[=covariate_step_tibble]{covariate_step_tibble()}}).}

\item{nm_col}{Character (default = "m"). Name of column to store nm objects.}

\item{parameters}{Character (default = "new").  Passed to \code{\link[=summary_wide]{summary_wide()}}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Extracts results from completed covariate runs and combines them into the covariate \code{\link[=tibble]{tibble()}}.
}
\examples{
\dontrun{
## create tibble of covariate step with model objects as column m
dsm1 <- m1 \%>\% covariate_step_tibble(
  run_id = "m1_f1",
  dtest = dtest,
  direction = "forward"
)

## run all models greedily
dsm1$m <- dsm1$m \%>\% run_nm()

wait_finish(dsm1$m)

## extract results and put into tibble
dsm1 <- dsm1 \%>\% bind_covariate_results()
}
}
\seealso{
\code{\link[=covariate_step_tibble]{covariate_step_tibble()}}
}
