% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{boot_to_csv}
\alias{boot_to_csv}
\title{save bootstrap datasets to disk}
\usage{
boot_to_csv(
  d,
  rsplit,
  data_name,
  data_folder = "DerivedData/bootstrap_datasets",
  overwrite = FALSE
)
}
\arguments{
\item{d}{dataset to merge}

\item{rsplit}{splits object from rsample}

\item{data_name}{name of dataset}

\item{data_folder}{path to bootstrap datasets}

\item{overwrite}{should datasets be overwritten}
}
\description{
intended to be used in a mutate statement from rsample::bootstrap output
}
