% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMproject.R
\name{build_modfile}
\alias{build_modfile}
\title{Document manual steps for traceability}
\usage{
build_modfile(code_section)
}
\arguments{
\item{code_section}{code section. A potential mix of R code and manual_edit() statements}
}
\value{
message to user
}
\description{
This function will not execute to prevent accidental partial execution
}
\examples{
\dontrun{
build_modfile({
  m16 \%>\% new_ctl("17") \%>\%  ## changes sdtab16 to sdtab17 and updated "based on: 16"
  update_parameters(m16) \%>\% ## update parameter with final estimates from run 16
  write_ctl() \%>\%            ## save results to file
  manual_edit("reparameterised CL -> K")  ## manual edit. Save afterwards
  commit_file("17")          ## (optional) snapshot file in version control system
})

## or equivalently:

build_modfile({
  m16 \%>\% new_ctl("17") \%>\%
  update_parameters(m16) \%>\% 
  write_ctl()
  manual_edit("17", "reparameterised CL -> K")
  commit_file("17")
})

}
}
