% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{change_parent}
\alias{change_parent}
\title{Change parent object}
\usage{
change_parent(m, mparent, silent = TRUE)
}
\arguments{
\item{m}{nm object}

\item{mparent}{new parent object to child}

\item{silent}{logical (default = TRUE). Should warn if conflicts detected}
}
\description{
Useful when create a \code{child()} of a modified run
}
