% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctl_handling.R
\name{change_seed}
\alias{change_seed}
\title{make new control file based on previous}
\usage{
change_seed(ctl, new_seed)
}
\arguments{
\item{ctl}{object of class nmexecute, ctl_list, ctl_character or character}

\item{new_seed}{character or numeric. new seed}
}
\value{
object of class ctl_character
}
\description{
make new control file based on previous
}
