% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_installation.R
\name{check_installation}
\alias{check_installation}
\title{Check NMproject installation}
\usage{
check_installation()
}
\value{
Logical \code{TRUE} or \code{FALSE} indicating whether the tests have succeeded
or not.  A test failure does not necessarily mean that NMproject
incorrectly configured.  Test messages will say whether they are needed or
just recommended to pass.
}
\description{
This function uses \code{testthat} to test whether pre-requisites are met.  It
first checks for availability of PsN via \code{system_nm()}, this is a requirement
for NMproject.  Subsequent checks are not mandatory, but recommended for full
functionality - these comprise a check to see if NMTRAN checks are configured
and a check to see if code completion has been set up.
}
