% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMproject.R
\name{check_session}
\alias{check_session}
\title{Check tidyproject for best practice compliance}
\usage{
check_session(proj_name = getwd(), silent = FALSE, check_rstudio = TRUE)
}
\arguments{
\item{proj_name}{character. default = current working directory. path to directory.}

\item{silent}{logical. default = FALSE. suppress messages or not}

\item{check_rstudio}{logical (default = FALSE). Check rstudio studio project directory}
}
\description{
Check tidyproject for best practice compliance
}

