% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_nm.R
\name{clean_run}
\alias{clean_run}
\title{Clean up temporary files}
\usage{
clean_run(m, output_loc = c("run_dir", "base"), include_grid_files = TRUE)
}
\arguments{
\item{m}{nm object}

\item{output_loc}{character either "run_dir" or "base}

\item{include_grid_files}{logical (default = TRUE) should slurm files be included}
}
\description{
Clean up temporary files
}
