% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{code_library}
\alias{code_library}
\title{Show Code Library}
\usage{
code_library(
  extn = NULL,
  fields = "Description",
  viewer = TRUE,
  silent = FALSE,
  return_info = FALSE
)
}
\arguments{
\item{extn}{vector string of extensions to include}

\item{fields}{character vector of fields to extract}

\item{viewer}{logical indicating if viewer should be used to display results
(default=FALSE)}

\item{silent}{logical indicating if messages should be silenced
(default=FALSE)}

\item{return_info}{logical (default = FALSE). Return data.frame of results
(FALSE= returns file paths)}
}
\description{
Show Code Library
}
\details{
requires \code{getOption("code_library_path")} to be set
}
\examples{

\dontrun{

code_library()

}
}
\seealso{
\code{\link{ls_code_library}}, \code{\link{preview}}
}
