% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_record.R
\name{coef.nm}
\alias{coef.nm}
\title{Gets parameters values}
\usage{
\method{coef}{nm}(object, ...)
}
\arguments{
\item{object}{object of class nmexecute}

\item{...}{trans argument (default = TRUE). logical.  Should parameters be transformed}
}
\value{
a \code{data.frame} of parameter values
}
\description{
Produces data.frame of parameter values
}

