% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_record.R
\name{coef_ext0}
\alias{coef_ext0}
\title{Gets parameters values}
\usage{
coef_ext0(ext.file)
}
\arguments{
\item{ext.file}{string of ext file}

\item{p_info}{parameter info object}
}
\value{
a \code{data.frame} of parameter values
}
\description{
Produces data.frame of parameter values
}

