% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_record.R
\name{coef_wide}
\alias{coef_wide}
\title{Coef (in wide format)}
\usage{
coef_wide(m, trans = TRUE)
}
\arguments{
\item{m}{nm object}

\item{trans}{logical (default = TRUE) should transform parameters?}
}
\description{
Coef (in wide format)
}
