% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-run-summaries.R
\name{coef_widelong}
\alias{coef_widelong}
\alias{coef_wide}
\alias{coef_long}
\title{Extract parameter values}
\usage{
coef_wide(m, trans = TRUE)

coef_long(m, trans = TRUE)
}
\arguments{
\item{m}{An nm object.}

\item{trans}{Logical (default = \code{TRUE}). Transform parameters using comments
in $THETA/$OMEGA/$SIGMA.}
}
\value{
\code{data.frame} of extracted model parameter values. \code{coef_wide()}
returns a \code{data.frame} in wide format. Vector valued objects \code{m}, will be
stacked vertically with one row per run.  \code{coef_long()} returns a
\code{data.frame} in long format. Vector valued objects \code{m}, will be stacked
horizontally.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Pulls parameters, standard errors, OFVs and condition numbers out of ext
files, applies transformations.  This function is useful when numeric values
are needed. \code{rr} is easier to read, however it returns characters.  A wide
and long format is available via two different functions.
}
\section{THETA transformations using \code{trans = TRUE}}{


Where FINAL is the reported estimate in the returned \code{tibble}, and RSE\% is
the reported standard error (where applicable) and \eqn{\theta} and \eqn{se(\theta)}
are the NONMEM reported values of parameters and standard errors, respectively

\describe{
\item{LOG or LOGODDS:}{
\eqn{FINAL = exp(\theta), RSE% = 100*\sqrt(exp^(se(\theta)) - 1)}
    }
\item{RATIO:}{
\eqn{FINAL = \theta, RSE% = 100*se(\theta) / \theta}
    }
\item{LOGIT:}{
\eqn{FINAL = 100 * 1 / (1 + exp(-\theta))}
}
}
}

\section{OMEGA transformations using \code{trans = TRUE}}{


Where FINAL is the reported estimate in the returned \code{tibble}, and RSE\% is
the reported standard error (where applicable) and \eqn{\omega^2} and \eqn{se(\omega^2)}
are the NONMEM reported values of parameters and standard errors, respectively

\describe{
\item{LOG:}{
\eqn{FINAL = 100*\sqrt(exp^(\omega^2) - 1), RSE% = 100*(se(\omega^2)/\omega^2)/2}
    }
}
}

\section{SIGMA transformations using \code{trans = TRUE}}{


Where FINAL is the reported estimate in the returned \code{tibble}, and RSE\% is
the reported standard error (where applicable) and \eqn{\sigma^2} and \eqn{se(\sigma^2)}
are the NONMEM reported values of parameters and standard errors, respectively

\describe{
\item{}{
\eqn{FINAL = \sqrt \omega^2, RSE% = 100*se(\omega^2) / \omega^2}
    }
}
}

\seealso{
\code{\link[=rr]{rr()}}
}
