% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{comment_out}
\alias{comment_out}
\title{comment lines of control file}
\usage{
comment_out(m, pattern = ".*")
}
\arguments{
\item{m}{nm object}

\item{pattern}{optional character regex.  Passed to gsub}
}
\description{
comment lines of control file
}
