% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_object.R
\name{completed_nm}
\alias{completed_nm}
\title{Create an nm object from an already completed PsN run}
\usage{
completed_nm(mod_file, data_path = data_name(mod_file))
}
\arguments{
\item{mod_file}{Path to a model file}

\item{data_path}{Optional path to the corresponding data set}
}
\value{
An nm object with class "completed_nm".
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This is useful if the run was made outside of NMproject and we want to
leverage NMproject to do postprocessing and results extraction.  It is not
recommended to apply \code{run_nm()} or \code{write_ctl()} to this object as that would
break reproducibility
}
