% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMproject.R
\name{copy_control}
\alias{copy_control}
\title{Copy NONMEM control stream}
\usage{
copy_control(from, to, overwrite = FALSE, alt_paths,
  dest_dir = getOption("models.dir"))
}
\arguments{
\item{from}{character. File to copy from}

\item{to}{character. File to copy to}

\item{overwrite}{logical. Should to file be overwritten? Default = FALSE.}

\item{alt_paths}{character vector. paths to other candidate files to search}

\item{dest_dir}{character. default "Models" dir}
}
\description{
Copy NONMEM control stream
}
