% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/developer_funs.R
\name{copy_demo_to_demo}
\alias{copy_demo_to_demo}
\title{Copy demo source files to package}
\usage{
copy_demo_to_demo(demo = "theopp")
}
\description{
Internal function (non exported)
  Use within a demo directory
}
