% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/developer_funs.R
\name{copy_demo_to_templates}
\alias{copy_demo_to_templates}
\title{Copy diagnostic funs to templates}
\usage{
copy_demo_to_templates()
}
\description{
Internal function (non exported)
  Use within a demo directory (e.g. after \code{run_all_scripts())})
}
