% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/developer_funs.R
\name{copy_demo_to_test}
\alias{copy_demo_to_test}
\title{Copy demo source files + relevant NM outputs to package testfiles}
\usage{
copy_demo_to_test()
}
\description{
Internal function (non exported)
  Use within a demo directory
}
