% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{copy_script2}
\alias{copy_script2}
\title{Copy script to project directory}
\usage{
copy_script2(
  from,
  to = file.path(scripts_dir(), basename(from)),
  stamp_copy = TRUE,
  overwrite = FALSE,
  comment_char = "#"
)
}
\arguments{
\item{from}{character. file name or path of file to copy}

\item{to}{character. file name.  default = same as from}

\item{stamp_copy}{logical. Create a commented timestamp at beginning of file}

\item{overwrite}{logical. Overwrite 'to' file if exists?}

\item{comment_char}{character. Comment character}
}
\description{
Will search code library and copy script and dependencies into scripts directory.
Script will also be stamped with source location, time and user information
}
