% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-run-summaries.R
\name{covariance_matrix}
\alias{covariance_matrix}
\alias{correlation_matrix}
\title{Get covariance & correlation matrices}
\usage{
covariance_matrix(m)

correlation_matrix(m)
}
\arguments{
\item{m}{An nm object.}
}
\value{
A \code{matrix} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Extracts covariance & correlation matrices from \code{nm} object.
}
\examples{

## requires NONMEM to be installed
\dontrun{

covariance_matrix(m)
correlation_matrix(m)

}
}
