% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-run-summaries.R
\name{covariance_plot}
\alias{covariance_plot}
\title{Plot $COV matrix}
\usage{
covariance_plot(r, trans = TRUE)
}
\arguments{
\item{r}{An nm object.}

\item{trans}{Logical (default = TRUE).  Applies the transformations specified
in $THETA/$OMEGA/$SIGMA comments before plotting.}
}
\value{
A \code{ggplot2} object with parameter correlations.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
Plots the correlation plot from the $COV NONMEM output.
}
\examples{

## requires NONMEM to be installed
\dontrun{

covariance_plot(m1)

}

}
\seealso{
\code{\link[=nm_render]{nm_render()}}
}
