% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{covariance_result}
\alias{covariance_result}
\title{Plot covariance matrix}
\usage{
covariance_result(r, trans = TRUE)
}
\arguments{
\item{r}{nm object}

\item{trans}{logical (default = TRUE)}
}
\description{
Plot covariance matrix
}
