% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_reestExtra.R
\name{covariate_step_tibble}
\alias{covariate_step_tibble}
\title{Prepare forward covariate step}
\usage{
covariate_step_tibble(base, run_id, run_in = getOption("models.dir"), dtest,
  direction = c("forward", "backward"), ...)
}
\arguments{
\item{base}{nm object}

\item{run_id}{run_id}

\item{run_in}{run_in}

\item{dtest}{tibble::tibble with testing relations}

\item{direction}{character. "forward" (default) or "backward"}

\item{...}{additional arguments passed to add_cov}
}
\value{
tibble::tibble dtest with appended columns
}
\description{
Prepare forward covariate step
}
