% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{ctl}
\alias{ctl}
\title{set/get control file object}
\usage{
ctl(m, ctl_ob, update_ctl = TRUE, ...)
}
\arguments{
\item{m}{nm object}

\item{ctl_ob}{optional path to control file}

\item{update_ctl}{logical. Should table numbers and author fields be updated}

\item{...}{additional arguments}
}
\description{
set/get control file object
}
