% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctl_handling.R
\name{ctl_character}
\alias{ctl_character}
\title{Constructor/converter to ctl_character}
\usage{
ctl_character(r)
}
\arguments{
\item{r}{either class nmexecute, character, ctl_list, ctl_character}
}
\value{
object of class ctl_character
}
\description{
Internal function
}
