% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctl_handling.R
\name{ctl_fix_eta}
\alias{ctl_fix_eta}
\title{fix etas}
\usage{
ctl_fix_eta(ctl, param, to)
}
\arguments{
\item{ctl}{object coercible into ctl_list}

\item{param}{character, name of paramater}

\item{to}{optional character/numeric, number to fix to}
}
\description{
fix etas
}
