% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctl_handling.R
\name{ctl_nm2r}
\alias{ctl_nm2r}
\title{Convert controls stream to first abstraction layer}
\usage{
ctl_nm2r(ls)
}
\arguments{
\item{ls}{character vector with NONMEM control stream contents}
}
\value{
object of class r.ctl
}
\description{
Convert controls stream to first abstraction layer
}
