% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctl_handling.R
\name{ctl_param_set}
\alias{ctl_param_set}
\title{set parameter value}
\usage{
ctl_param_set(ctl, param, to)
}
\arguments{
\item{ctl}{object coercible into ctl_list}

\item{param}{character, name of paramater}

\item{to}{character/numeric, number to fix to}
}
\description{
set parameter value
}
