% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-gettersetters.R
\name{ctl_path}
\alias{ctl_path}
\title{Get and set path to NONMEM control file}
\usage{
ctl_path(m, text)
}
\arguments{
\item{m}{An nm object.}

\item{text}{Optional character. Name of path to control file (see details).
Typically, this file does not yet normally exist, but will house the code
code for this run.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Similar to \code{\link[=ctl_name]{ctl_name()}} & \code{\link[=run_in]{run_in()}}, this allows you to retrieve and specify
the relative path to the control file that will be written by the \code{\link[=run_nm]{run_nm()}}.
}
\details{
Note that \code{text} can contain an \code{"{run_id}"} string.  E.g.
\code{"Models/run{run_id}.mod"} will use the name \code{"Models/runm1.mod"} if
\code{run_id(m1)} is \code{"m1"}.
}
\examples{
\dontrun{

m0 <- nm(run_id = "m0")
ctl_name(m0)

m0 <- m0 \%>\% ctl_name("run{run_id}.mod")
ctl_name(m0)

## warning: this makes the run identifier in old xpose4 "_m0" not "m0"!
m0 <- m0 \%>\% ctl_path("Models/run_{run_id}.mod")
ctl_path(m0)
}
}
