% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctl_handling.R
\name{ctl_r2nm}
\alias{ctl_r2nm}
\title{Convert first abstraction layer to control stream}
\usage{
ctl_r2nm(x)
}
\arguments{
\item{x}{object of class r.ctl}
}
\value{
character vector with control stream contents
}
\description{
Convert first abstraction layer to control stream
}

