% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctl_handling.R
\name{ctl_unfix_eta}
\alias{ctl_unfix_eta}
\title{unfix etas}
\usage{
ctl_unfix_eta(ctl, param)
}
\arguments{
\item{ctl}{object coercible into ctl_list}

\item{param}{character, name of paramater}
}
\description{
unfix etas
}
