% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{data_filter_char}
\alias{data_filter_char}
\title{Get filter statement}
\usage{
data_filter_char(r, ...)
}
\arguments{
\item{r}{object coercible into ctl_list}

\item{...}{arguments passed to data_ignore_char}
}
\description{
Opposite of data_ignore_char
}
