% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{data_ignore_char}
\alias{data_ignore_char}
\title{Get ignore statement}
\usage{
data_ignore_char(r, data)
}
\arguments{
\item{r}{object coercible into ctl_list}

\item{data}{data.frame (default = missing) optional input dataset from r}
}
\description{
Get ignore statement
}
