% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMproject.R
\name{data_name.default}
\alias{data_name.default}
\title{Get NONMEM dataset name from control stream}
\usage{
\method{data_name}{default}(x)
}
\arguments{
\item{x}{character vector. Name(s) of control stream(s)}
}
\description{
Get NONMEM dataset name from control stream
}

