% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMproject.R
\name{data_name}
\alias{data_name}
\title{Get NONMEM dataset name from control stream}
\usage{
data_name(x)
}
\arguments{
\item{x}{argument to be passed to method.}
}
\description{
Get NONMEM dataset name from control stream
}

