% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{data_path}
\alias{data_path}
\title{Get/set path to dataset}
\usage{
data_path(m, text)
}
\arguments{
\item{m}{nm object}

\item{text}{optional character. Path to input dataset}
}
\description{
Get/set path to dataset
}
