% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{delete_nm}
\alias{delete_nm}
\title{Delete run from database}
\usage{
delete_nm(db_name = "runs.sqlite", entry)
}
\arguments{
\item{db_name}{character. Name of db}

\item{entry}{numeric. entry name to delete}
}
\description{
Delete run from database
}
