% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dollar_data.R
\name{dollar_data}
\alias{dollar_data}
\title{Generate $INPUT values from NONMEM compatible dataset}
\usage{
dollar_data(d, keep, rename)
}
\arguments{
\item{d}{data.frame. NONMEM compatible dataset}

\item{keep}{character vector. Names of columns to keep}

\item{rename}{named character vector. Renaming instructions}
}
\description{
Generate $INPUT values from NONMEM compatible dataset
}
