% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{exclude_rows}
\alias{exclude_rows}
\title{Exclude rows of NONMEM dataset}
\usage{
exclude_rows(d, dexcl, exclude_col = "EXCL")
}
\arguments{
\item{d}{data.frame for NONMEM dataset}

\item{dexcl}{data.frame consisting of rows to be ignored}

\item{exclude_col}{character.  Name of exclude column in d}
}
\description{
Exclude rows of NONMEM dataset
}
\examples{
\dontrun{
## use with dplyr
dexcl <- d \%>\% filter(ID == 23, TIME > 18, TIME < 24) \%>\% select(ID, TIME, DV, EXCL)
dexcl  ## view rows to be excluded
d <- d \%>\% exclude_rows(dexcl)
}
}
