% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{extra_files}
\alias{extra_files}
\title{List files to be cleaned up}
\usage{
extra_files(r)
}
\arguments{
\item{r}{object class nm}
}
\description{
List files to be cleaned up
}
\examples{
\dontrun{
mod1 \%>\% extra_files
mod1 \%>\% extra_files \%>\% cleanup
}
}
