% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{extract_nm}
\alias{extract_nm}
\title{Extract nm object from data.base}
\usage{
extract_nm(entry, db_name = "runs.sqlite")
}
\arguments{
\item{entry}{numeric. Database ID of nm object}

\item{db_name}{character. Name of db}
}
\description{
Extract nm object from data.base
}
