% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMproject.R
\name{find_nm_install_path}
\alias{find_nm_install_path}
\title{Find location of NONMEM}
\usage{
find_nm_install_path(name = "default")
}
\arguments{
\item{name}{character name of nonmem installation (according PsN)}
}
\description{
Find location of NONMEM
}
