% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nonmem.R
\name{find_nm_tran_path}
\alias{find_nm_tran_path}
\title{Find location of NMTRAN}
\usage{
find_nm_tran_path(name = "default", warn = TRUE)
}
\arguments{
\item{name}{character name of nonmem installation (according PsN)}

\item{warn}{logical (default = TRUE), should warning be given if fail to find
NMTRAN.exe}
}
\value{
If cannot find installation will return \code{NULL} without errors. Will
emit warnings \code{warm = TRUE}.
}
\description{
Attempts to find location of NMTRAN used by PsN.
}
\details{
The function will attempt to use a locally available PsN
installation to get this information.  If the PsN installation is on a
remote server, this function will not work (it will return a \code{NULL}). If
cannot find installation, you will need to set
\code{\link[=nm_tran_command]{nm_tran_command()}}, manually
}
\seealso{
\code{\link[=nm_tran_command]{nm_tran_command()}}
}
