% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_gettersetters.R
\name{find_result_files}
\alias{find_result_files}
\title{get paths of results files}
\usage{
find_result_files(r, pattern = ".")
}
\arguments{
\item{r}{nm object}

\item{pattern}{grep matching pattern}
}
\description{
get paths of results files
}
