% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise_covariate.R
\name{gen_sim_path}
\alias{gen_sim_path}
\title{Generate paths for simulation runs}
\usage{
gen_sim_path(dsc, start_dir, include_names = TRUE)
}
\arguments{
\item{dsc}{\code{dplyr::tibble} with experimental factors to vary}

\item{start_dir}{directory name within which all runs should take place.}

\item{include_names}{logical (default = TRUE).}
}
\description{
Internal function
}
\examples{
\dontrun{

dsc$location <- gen_sim_path(dsc, ...)

}
}
