% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctl_handling.R
\name{get_data}
\alias{get_data}
\title{get data set}
\usage{
get_data(r, ...)
}
\arguments{
\item{r}{object coercible into ctl_character}

\item{...}{additional arguments for read.csv}
}
\description{
get data set
}
