% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{git_cmd_available}
\alias{git_cmd_available}
\title{Check if git is available on command line}
\format{
An object of class \code{logical} of length 1.
}
\usage{
git_cmd_available
}
\description{
Check if git is available on command line
}
\keyword{internal}
