% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{git_hooks}
\alias{git_hooks}
\alias{nm_pre_commit_hook}
\alias{nm_pre_push_hook}
\title{Git hooks}
\usage{
nm_pre_commit_hook(cmd)

nm_pre_push_hook(cmd)
}
\arguments{
\item{cmd}{Optional command for setting the git hook.}
}
\value{
If no \code{cmd} is specified this returns the return value of
\code{getOption("nm_pre_commit_hook")}.  Otherwise there is no return value.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function is primarily for organisational level configuration. Supply
git hooks in the form of R functions and they will be executed via the
\code{Rscript} interface.
}
\seealso{
\code{\link[=nm_create_analysis_project]{nm_create_analysis_project()}}
}
\keyword{internal}
