% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%f>\%}
\alias{\%f>\%}
\title{Function pipe for nm objects}
\usage{
lhs \%f>\% rhs
}
\arguments{
\item{lhs}{An nm object.}

\item{rhs}{A list of functions.  Must be same length as \code{lhs}.}
}
\value{
A modified nm object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Pipe an nm object object to a list of functions.  Although this enables
multiple NONMEM runs to be handled simulataneously, it does make your code
less readable.
}
\examples{

\dontrun{
mWT <- m2 \%>\% child(c("m3", "m4", "m5", "m6", "m7", "m8")) \%f>\% 
list(
 . \%>\% add_cov(param = "CL", cov = "WT", state = "power"),
 . \%>\% add_cov(param = "CL", cov = "WT", state = "power0.75"),
 . \%>\% add_cov(param = "V", cov = "WT", state = "power"),
 . \%>\% add_cov(param = "V", cov = "WT", state = "power1"),
 . \%>\% add_cov(param = "CL", cov = "WT", state = "power") \%>\%
   add_cov(param = "V", cov = "WT", state = "power"),
 . \%>\% add_cov(param = "CL", cov = "WT", state = "power0.75") \%>\%
  add_cov(param = "V", cov = "WT", state = "power1")
)

mWT \%>\% run_nm() \%>\% wait_finish()

mWT \%>\% summary_wide()

}

}
\seealso{
\code{\link[=child]{child()}} for creating multiple child NONMEM objects
}
