% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctl_handling.R
\name{gsub_ctl}
\alias{gsub_ctl}
\title{Pattern replacement for ctl file}
\usage{
gsub_ctl(ctl, pattern, replacement, ..., dollar = NA_character_)
}
\arguments{
\item{ctl}{object coercible into ctl_character}

\item{pattern}{argument passed to gsub}

\item{replacement}{argument passed to gsub}

\item{...}{additional arguments passed to gsub}

\item{dollar}{character name of subroutine}
}
\description{
A wrapper around \code{gsub} so that control files may be modified using
\code{gsub} syntax.
}
