% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{info_scripts}
\alias{info_scripts}
\title{List information about scripts}
\usage{
info_scripts(
  files,
  fields = c("Description"),
  viewer = TRUE,
  silent = FALSE,
  base_dirs = NULL,
  shorten_paths = TRUE
)
}
\arguments{
\item{files}{vector string of file names/paths}

\item{fields}{vector string of field tags to display}

\item{viewer}{logical indicating if Rstudio viewer should be used (default = TRUE)}

\item{silent}{run in quiet mode (default=FALSE)}

\item{base_dirs}{character vector. group files together that belong to these directory paths}

\item{shorten_paths}{logical. Default = TRUE. Long paths will be shortened if true in displayed output (not returned object)}
}
\description{
List information about scripts
}
\examples{
\dontrun{
ls_scripts('~/AZD6094/PK_liver4/') \%>\%
  info_scripts('Description') \%>\%
  filter(grepl('mod',DESCRIPTION))
}
}
