% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_object.R
\name{input_data}
\alias{input_data}
\title{Read input dataset}
\usage{
input_data(m, filter = FALSE, na = ".", ...)
}
\arguments{
\item{m}{nm object}

\item{filter}{logical (default = FALSE). should NONMEM ignore statement be applied}

\item{na}{character. passed to read.csv}

\item{...}{additional arguments passed to either read_derived_data or read.csv}
}
\description{
Read input dataset
}
