% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{interactive_mode}
\alias{interactive_mode}
\title{Should run() work in interactive mode}
\usage{
interactive_mode(value)
}
\arguments{
\item{value}{logical. TRUE = use interactive mode. FALSE = use non-interactive mode}
}
\description{
Should run() work in interactive mode
}
