% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring.R
\name{is_finished}
\alias{is_finished}
\title{Tests if job is finished}
\usage{
is_finished(r, initial_timeout = NA)
}
\arguments{
\item{r}{object class nm}

\item{initial_timeout}{numeric. time in seconds.
time period to give up on a run if directory hasn't been created.}
}
\description{
Tests if job is finished
}
