% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsExtra.R
\name{is_nm}
\alias{is_nm}
\alias{is_nm_list}
\alias{is_nm_generic}
\title{Test if object is an nm coercible object}
\usage{
is_nm_list(x)

is_nm_generic(x)
}
\arguments{
\item{x}{Object.}
}
\description{
Mostly internal functions.
}
