% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMproject.R
\name{is_nm_file_name}
\alias{is_nm_file_name}
\title{test if NONMEM file name conforms to convention}
\usage{
is_nm_file_name(x, error_if_false = FALSE)
}
\arguments{
\item{x}{character vector. file name or path}

\item{error_if_false}{logical. Default=FALSE. If true, will make an error if test fails}
}
\description{
test if NONMEM file name conforms to convention
}
