% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-management.R
\name{is_nmproject_dir}
\alias{is_nmproject_dir}
\title{Is the directory an NMproject directory}
\usage{
is_nmproject_dir(path = getwd())
}
\arguments{
\item{path}{Optional path to test if it's an NMproject or not.}
}
\value{
Logical \code{TRUE} or \code{FALSE}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Find out whether current (or specified) directory is an NMproject directory
or not.
}
