% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_rstudio}
\alias{is_rstudio}
\title{Logical flag for detecting if R session is on RStudio}
\usage{
is_rstudio()
}
\value{
A logical \code{TRUE} or \code{FALSE} value.
}
\description{
Logical flag for detecting if R session is on RStudio
}
\keyword{internal}
