% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{is_status_finished}
\alias{is_status_finished}
\title{tests if job is finished}
\usage{
is_status_finished(status_ob)
}
\arguments{
\item{status_ob}{output from run_status}
}
\description{
tests if job is finished
}
