% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMproject-options.R
\name{job_time_spacing}
\alias{job_time_spacing}
\title{Setup default job_time_spacing option}
\usage{
job_time_spacing(seconds)
}
\arguments{
\item{seconds}{Optional numeric value. Values correspond to what will be set.}
}
\value{
if \code{seconds} is missing, will return value of
\code{getOption("job_time_spacing")} otherwise will set option \code{job_time_spacing}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This allows organisations/individuals with their own job time spacing, used by \code{\link[=run_nm]{run_nm()}} when \code{threads > 1}.
}
\examples{

job_time_spacing()
job_time_spacing(1)

}
