% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system-hooks.R
\name{kill_job}
\alias{kill_job}
\title{Kill cluster job}
\usage{
kill_job(m)
}
\arguments{
\item{m}{An nm object.}
}
\description{
Can be useful in conjuction with a modified \code{\link[=system_nm]{system_nm()}}.  Requires setting \code{"kill_job"} option.
}
\details{
The option \code{"kill_job"} needs to be defined as a function that kills
a cluster job.
}
\seealso{
\code{\link[=system_nm]{system_nm()}}, \code{\link[=job_info]{job_info()}}
}
\keyword{internal}
