% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_dirs}
\alias{list_dirs}
\title{List directories}
\usage{
list_dirs(
  path = ".",
  full.names = TRUE,
  recursive = FALSE,
  maxdepth = 1,
  pattern
)
}
\arguments{
\item{path}{Same as \code{list.dirs()}.}

\item{full.names}{Same as \code{list.dirs()}.}

\item{recursive}{Same as \code{list.dirs()}.}

\item{maxdepth}{Integer (default = \code{1}) maximum depth to search.}

\item{pattern}{Optional character (default is missing) regex pattern match on directory
name.}
}
\value{
Character vector of paths to matched directories.
}
\description{
Wrapper around \code{list.dirs()} but includes \code{maxdepth} and pattern arguments and
removes \code{full.names} argument, always return full names.
}
\keyword{internal}
