% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sse.R
\name{load_sse}
\alias{load_sse}
\title{Load SSE data}
\usage{
load_sse(x, parr = F, nodes = 1)
}
\arguments{
\item{x}{object class nm}

\item{parr}{logical. include description here#'}

\item{nodes}{include description here}
}
\description{
Function to load simulation data from PsN SSE
This facilitates extracting data needed for postprocessing of simulations
poptentially helpfor after simulation with uncertainty (e.g. obtained through PsN bootstrap or impance sampling)
}
