% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMproject.R
\name{load_top_level_models}
\alias{load_top_level_models}
\title{load all model objects defined in a script}
\usage{
load_top_level_models(script_name)
}
\arguments{
\item{script_name}{path to script file}
}
\description{
Will rerun all '<- nm()' definitions
Only remakes top level definitions.
Programmatic use of nm() (e.g. in a for loop or using paste with other objects) will generally not work.
Sorry!
}
