% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-code.R
\name{ls_code_library}
\alias{ls_code_library}
\title{List files in code library}
\usage{
ls_code_library(pattern = ".")
}
\arguments{
\item{pattern}{optional character. filter the code library use regex}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

A low level function to interact with the code library.  It is easier in most
cases to use the shiny "code library" RStudio 'Addin'.
}
\examples{

\dontrun{

ls_code_library("Models/ADVAN2.mod") \%>\%
  stage()
}

}
\seealso{
\code{\link[=code_library]{code_library()}}, \code{\link[=preview]{preview()}}, \code{\link[=stage]{stage()}}, \code{\link[=import]{import()}}
}
