% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{ls_scripts}
\alias{ls_scripts}
\title{List scripts}
\usage{
ls_scripts(folder = ".", extn = "r|R|Rmd|rmd", recursive = TRUE)
}
\arguments{
\item{folder}{string describing folder to search recursively in}

\item{extn}{character (can be regex) giving extension to limit search to}

\item{recursive}{by default TRUE}
}
\description{
List scripts
}
\examples{
\dontrun{
ls_scripts('~/AZD6094/PK_liver4/') \%>\%
  info_scripts('Description') \%>\%
  filter(grepl('mod',DESCRIPTION))
}
}
