% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{ls_tempfiles_old}
\alias{ls_tempfiles_old}
\title{get all temp files}
\usage{
ls_tempfiles_old(
  object = ".",
  output_loc = c("run_dir", "base"),
  run_files = NA_character_,
  include_grid_files = TRUE
)
}
\arguments{
\item{object}{nm object or path to project (default = ".")}

\item{output_loc}{character either "run_dir" or "base}

\item{run_files}{optional character with NM_run* file paths}

\item{include_grid_files}{logical (default = TRUE) should slurm files be included}
}
\description{
list all tempfiles (normally for deletion)
}
