% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{make_boot_df}
\alias{make_boot_df}
\title{Write bootstrap dataset}
\usage{
make_boot_df(m, replicates, dboot_rds)
}
\arguments{
\item{m}{nm object}

\item{replicates}{numeric vector of bootstrap replicate numbers}

\item{dboot_rds}{path to bootstrap RDS file}
}
\description{
Write bootstrap dataset
}
