% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{make_xv_datasets}
\alias{make_xv_datasets}
\title{add/write xv datasets}
\usage{
make_xv_datasets(
  dboot,
  data_folder = "DerivedData/bootstrap_datasets",
  overwrite = FALSE,
  id_var = "ID"
)
}
\arguments{
\item{dboot}{output from make_boot_dataset()}

\item{data_folder}{folder to store datasets}

\item{overwrite}{overwrite or not}

\item{id_var}{character (default = "ID"). Name of ID column}
}
\description{
add/write xv datasets
}
