% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMproject.R
\name{manual}
\alias{manual}
\title{Document manual steps for traceability}
\usage{
manual(code_section)
}
\arguments{
\item{code_section}{code section. A potential mix of R code and manual_step() statements}
}
\value{
message to user
}
\description{
This function will not execute to prevent accidental partial execution
}
\examples{
\dontrun{
manual({
  m16 \%>\% update_parameters(m16) \%>\% write_ctl("17")
  manual_step("reparameterised CL -> K")
})
}
}
