% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMproject.R
\name{manual_edit}
\alias{manual_edit}
\title{Document manual edit}
\usage{
manual_edit(comment)
}
\arguments{
\item{comment}{character. Description of change}
}
\value{
error with comment name
}
\description{
Document manual edit
}
