% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMproject.R
\name{manual_step}
\alias{manual_step}
\title{Document manual step}
\usage{
manual_step(comment)
}
\arguments{
\item{comment}{character. Description of change}
}
\value{
error with comment name
}
\description{
Document manual step
}
