% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctl_handling.R
\name{model_file_name}
\alias{model_file_name}
\title{Generate model file name}
\usage{
model_file_name(run_id, dir = getOption("models.dir"), mustWork = FALSE)
}
\arguments{
\item{run_id}{run identifier}

\item{dir}{directory to work in}

\item{mustWork}{same as normalizePath argument}
}
\description{
Generate model file name
}
