% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMproject.R
\name{new_jobs}
\alias{new_jobs}
\title{get/set new_jobs}
\usage{
new_jobs(txt = c("ask", "overwrite", "stop new", "skip"))
}
\arguments{
\item{txt}{character either "run", "stop", "skip"}
}
\description{
Requires setting "new_jobs" \code{option()}
}
