% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_object.R
\name{new_nm}
\alias{new_nm}
\title{create a new nm object}
\usage{
new_nm(
  run_id = NA_character_,
  based_on,
  data_path,
  cmd = "execute {ctl_name} -dir={run_dir}"
)
}
\arguments{
\item{run_id}{character. Run identifier}

\item{based_on}{character. Prior ctl file to base run on}

\item{data_path}{character. Path to dataset}

\item{cmd}{character. Psn command to use}
}
\description{
Wrapper function around nm()
}
\examples{
\dontrun{
m1 <- new_nm(run_id = "m1",
            based_on = "staging/Models/run1.mod",
            data_path = "DerivedData/data.csv",
            cmd = "execute -run_on_sge -sge_prepend_flags='-V' {ctl_name} -dir={run_dir}")
}
}
