% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{new_notebook_template}
\alias{new_notebook_template}
\title{Create new R notebook}
\usage{
new_notebook_template(script_name, overwrite = FALSE, open_file = TRUE,
  libs = c("NMproject"))
}
\arguments{
\item{script_name}{character}

\item{overwrite}{logical. Whether to overwrite existing file (default = FALSE)}

\item{open_file}{logical. Whether function should open script (default = TRUE)}

\item{libs}{character. What libraries to add.}
}
\description{
Create new R notebook
}
