% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMproject.R
\name{new_script}
\alias{new_script}
\title{Create new R script}
\usage{
new_script(name, overwrite = FALSE, open_file = TRUE,
  libs = c("NMproject"))
}
\arguments{
\item{name}{character indicating name of script to create}

\item{overwrite}{logical. Whether to overwrite existing file (default = FALSE)}

\item{open_file}{logical. Whether function should open script (default = TRUE)}

\item{libs}{character. What libraries to add.}
}
\description{
Create new R script
}
