% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{nm}
\alias{nm}
\title{Make object of class nm}
\usage{
nm(cmd, parent = NULL, description = "", psn_command = NULL,
  shell_script_name = NULL, shell_script_extn = "sh", ctl_name = NULL,
  run_id = NULL, run_dir = NULL, run_in = getOption("models.dir"),
  db_name = "runs.sqlite", scm_config_name = NULL)
}
\arguments{
\item{cmd}{character. system command to launch NONMEM (PsN)}

\item{parent}{object of class nm. Parent object to create (optional)}

\item{description}{character. Description of model (optional)}

\item{psn_command}{character. Name of PsN command (optional)}

\item{shell_script_name}{character. Name of shell script for grid submission (optional)}

\item{shell_script_extn}{character. File extension shell script for grid submission (optional)}

\item{ctl_name}{character. Name of control file (optional)}

\item{run_id}{character or numeric. Run identifier (optional)}

\item{run_dir}{character or numeric. Run directory (optional)}

\item{run_in}{character. Directory to execute. Default = getOption("models.dir")}

\item{db_name}{character. Name of db. NULL = don't put in database}

\item{scm_config_name}{character. Name of scm config file (optional)}
}
\value{
object of class nm
}
\description{
Make object of class nm
}
