% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{nm_boot}
\alias{nm_boot}
\title{Write bootstrap dataset}
\usage{
nm_boot(m, replicates, dboot_rds = "dboots_big.RDS")
}
\arguments{
\item{m}{nm object}

\item{replicates}{numeric vector of bootstrap replicate numbers}

\item{dboot_rds}{path to bootstrap RDS file}
}
\description{
Write bootstrap dataset
}
