% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{nm_create_analysis_project}
\alias{nm_create_analysis_project}
\title{Create analysis project}
\usage{
nm_create_analysis_project(
  path,
  dirs = nm_default_dirs(),
  style = c("simple"),
  use_renv = FALSE,
  ...
)
}
\arguments{
\item{path}{character path (relative or absolute) to project.  If just
specifying a name, this will create the analysis project in the current
working directory}

\item{dirs}{character list or vector.  Default = \code{nm_default_dirs()}}

\item{style}{character. Currently only "simple" See details}

\item{use_renv}{logical (default = FALSE). should renv be used or not in
project.}

\item{...}{deprecated}
}
\description{
Easiest way to use this function is via RStudio menus (FILE -> New Project ->
New Directory -> New NMproject).  However the function can be used directly
too.
}
\details{
The function works like as is inspired by
\code{starters::create_analysis_project}. There is no restriction on directory
name.  It is therefore possible to violate R package naming conventions.
This is to cater to users who like underscores and aren't interested in
creating a package.
}
