% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMproject.R
\name{nm_default_dir}
\alias{nm_default_dir}
\title{Get a default directory}
\usage{
nm_default_dir(name = c("scripts", "models", "results"), ...)
}
\arguments{
\item{name}{character. Directory type}

\item{...}{not used yet}
}
\description{
Get a default directory
}
\examples{
nm_default_dir("scripts")
nm_default_dir("models")
nm_default_dir("results")
}
