% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMproject-options.R
\name{nm_default_fields}
\alias{nm_default_fields}
\title{Setup default nm object fields}
\usage{
nm_default_fields(field_list)
}
\arguments{
\item{field_list}{Optional named list or vector. Names correspond to function
names and object fields, values correspond to what will be set.}
}
\value{
if \code{field_list} is missing, will return value of
\code{getOption("nm_default_fields")} otherwise will set option \code{nm_default_fields}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This allows organisations/individuals with their own nm object field
preferences to set these.
}
\examples{

nm_default_fields()
nm_default_fields(list(
  cmd = "execute {ctl_name} -dir={run_dir}"
))
nm_default_fields()

}
