% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMproject.R
\name{nm_dirs}
\alias{nm_dirs}
\title{setup analysis subdirectories}
\usage{
nm_dirs(dir_list)
}
\arguments{
\item{dir_list}{optional named list or vector. Names "scripts_dir" and "models_dir" must be present.  Rest can be unnamed}
}
\value{
if `dir_list` is missing, will return value of `getOption("nm_dirs")` otherwise will set option `nm_dirs`
}
\description{
This allows organisations/individuals with their own directory to customize their directory structure
}
\examples{
\dontrun{

nm_dirs()
nm_dirs(list(models_dir = "Models", 
             scripts_dir = "Scripts",
             results_dir = "Results",
             "SourceData",
             "Data"))
}
}
