% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_object.R
\name{nm_list_gather}
\alias{nm_list_gather}
\title{Get all nm_list objects}
\usage{
nm_list_gather(x = .GlobalEnv)
}
\arguments{
\item{x}{environment (default = .GlobalEnv) to search
or data.frame with (nm_list column) or nm_list}
}
\description{
Get all nm_list objects
}
