% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{nm_output}
\alias{nm_output}
\title{Get NONMEM output tables}
\usage{
nm_output(r, dorig, ...)
}
\arguments{
\item{r}{An object of class \code{nmexecute}}

\item{dorig}{Optional data.frame. NONMEM input dataset.}

\item{...}{Additional arguments to pass on to \code{\link[=read.csv]{read.csv()}}.}
}
\description{
This combines $TABLE output with the input data, allowing text columns to be retained for plotting/summaries.
}
