% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.table.nm.R
\name{nm_read_table}
\alias{nm_read_table}
\title{Fast read of NONMEM output table}
\usage{
nm_read_table(file, ...)
}
\arguments{
\item{file}{file argument - see ?read.table help for details}

\item{...}{other arguments to be passed to read.table()}
}
\description{
Fast read of NONMEM output table
}
