% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_render.R
\name{nm_render}
\alias{nm_render}
\title{render function for nm objects}
\usage{
nm_render(
  m,
  input,
  output_file = NA,
  args = list(),
  force = getOption("nm.force_render"),
  async = FALSE,
  ...
)
}
\arguments{
\item{m}{nm object}

\item{input}{character. Same as rmarkdown::render() arg}

\item{output_file}{character. Same as rmarkdown::render() arg}

\item{args}{list. Same as "params" arg in rmarkdown::render()}

\item{force}{logical (default = \code{getOption("nm.force_render")}). will force execution}

\item{async}{experimental option to use future package}

\item{...}{additional argument passed to rmarkdown::render()}
}
\description{
render function for nm objects
}
\details{
\code{input} must refer to a properly specified Rmd document.
The R markdown template "model diagnostic" in RStudio sets this up 
for you.

These R markdown templates are usable as R Notebooks (e.g. for code
development and debugging) if the object \code{.m} is defined in the
global work space first.
}
\examples{
\dontrun{
  m1 \%>\% nm_render("Scripts/basic_gof.Rmd")

  ## to run "Scripts/basic_gof.Rmd" as an R Notebook
  ## first define .m
  
  .m <- m1 ## Now you can run "Scripts/basic_gof.Rmd" as a Notebook
}
}
