% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMproject.R
\name{nm_tran.default}
\alias{nm_tran.default}
\title{Run NMTRAN step only}
\usage{
\method{nm_tran}{default}(x)
}
\arguments{
\item{x}{character. name/path of model file}
}
\description{
Run NMTRAN step only
}

