% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMproject.R
\name{nm_tran}
\alias{nm_tran}
\title{Run NMTRAN step only}
\usage{
nm_tran(x)
}
\arguments{
\item{x}{character. file name of NONMEM control stream}
}
\description{
Requires options('path/to/nmtran') to be set up.
}

