% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMproject.R
\name{nm_tran_command}
\alias{nm_tran_command}
\title{get/set nm_tran_command}
\usage{
nm_tran_command(text)
}
\arguments{
\item{text}{optional character. If specified will set nm_tran_command}
}
\description{
get/set nm_tran_command
}
