% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{nm_tree}
\alias{nm_tree}
\title{Make data.tree object}
\usage{
nm_tree(..., summary = FALSE)
}
\arguments{
\item{...}{arguments passed to nm_list_gather(...)}

\item{summary}{logical (default = FALSE) should summary_wide variables be appended}
}
\description{
Make data.tree object
}
