% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{nmdb_extract}
\alias{nmdb_extract}
\title{Extract nm object from data.base}
\usage{
nmdb_extract(entry)
}
\arguments{
\item{entry}{numeric. Database ID of nm object}
}
\description{
Extract nm object from data.base
}

