% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMproject.R
\docType{package}
\name{NMproject}
\alias{NMproject}
\title{Script Based 'NONMEM' Model Development}
\description{
Industrialisation of 'NONMEM' model development via end-to-end model
development workflows.
}
\section{Getting started}{


  URL: \url{http://tsahota.github.io/NMproject/}

  The primary source of documentation is the webpage. The vignette is on the
  "Getting Started" and the demo is accessible via the
  \code{\link{setup_nm_demo}} function.  To create a new project to work
  within see \code{\link{nm_create_analysis_project}}.
}

\seealso{
\code{\link{setup_nm_demo}},
  \code{\link{nm_create_analysis_project}}
}
