% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{nmsave_plot}
\alias{nmsave_plot}
\title{Save plots in results_dir}
\usage{
nmsave_plot(
  r,
  plot_ob,
  plot_name,
  plot_dir = results_dir(r),
  width = 7,
  height = 5,
  dpi = 300,
  ...
)
}
\arguments{
\item{r}{nm object}

\item{plot_ob}{a list of plotting objects}

\item{plot_name}{character. Name of results file}

\item{plot_dir}{character (default = results_dir(r)). Where to save.}

\item{width}{passed to ggsave}

\item{height}{passed to ggsave}

\item{dpi}{passed to ggsave}

\item{...}{passed to ggsave}
}
\description{
Save plots in results_dir
}
