% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmsave.R
\name{nmsave_table}
\alias{nmsave_table}
\title{Save plots in results_dir}
\usage{
nmsave_table(r, table_ob, table_name, table_dir = results_dir(r), ...)
}
\arguments{
\item{r}{nm object}

\item{table_ob}{a list of table objects}

\item{table_name}{character. Name of results file}

\item{table_dir}{character (default = results_dir(r)). Where to save.}

\item{...}{passed to ggsave}
}
\description{
Save plots in results_dir
}
