% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMproject.R
\name{NONMEM_version}
\alias{NONMEM_version}
\title{Get NONMEM version info}
\usage{
NONMEM_version()
}
\value{
returns list with version info for NONMEM, PsN, perl and
fortran compiler (only gfortran currently)
}
\description{
Get NONMEM version info
}
