% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctl_handling.R
\name{ofv}
\alias{ofv}
\title{Get objective function value (OFV)}
\usage{
ofv(r)
}
\arguments{
\item{r}{object of class nm}
}
\description{
Get objective function value (OFV)
}
\examples{
\dontrun{

ofv(m1)

}
}
