% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-run-summaries.R
\name{omega_matrix}
\alias{omega_matrix}
\title{Get OMEGA matrix from run}
\usage{
omega_matrix(r)
}
\arguments{
\item{r}{An nm object.}
}
\value{
A \code{matrix} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Obtain in matrix form the OMEGA matrix.  This is primarily to feed into other
packages such as \code{mrgsolve}.
}
\examples{

## requires NONMEM to be installed
\dontrun{

## matrix of initial estimates
m1 \%>\% omega_matrix()

## matrix of final estimates
m1 \%>\%
  update_parameters() \%>\%
  omega_matrix()
}

}
