% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{output_table}
\alias{output_table}
\title{Get processed output table}
\usage{
output_table(r, ...)
}
\arguments{
\item{r}{object of class nm}

\item{...}{optional additional arguments to pass on to read.csv of orig data}
}
\description{
Get processed output table
}
