% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{overlapping_outputs}
\alias{overlapping_outputs}
\title{Determine if runs are expected to have overlapping output files (from PsN)}
\usage{
overlapping_outputs(m)
}
\arguments{
\item{m}{nm_list object}
}
\description{
Determine if runs are expected to have overlapping output files (from PsN)
}
