% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{overwrite_default}
\alias{overwrite_default}
\title{Should run() overwrite previously run jobs}
\usage{
overwrite_default(x)
}
\arguments{
\item{x}{logical. TRUE means run() will overwrite previous runs by default,}
}
\description{
Should run() overwrite previously run jobs
}

