% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate-explore.R
\name{param_cov_diag}
\alias{param_cov_diag}
\title{Plot relationship between a parameter and covariate}
\usage{
param_cov_diag(r, param, cov, ..., categorical = FALSE, plot_tv = TRUE)
}
\arguments{
\item{r}{An nm object.}

\item{param}{Character. Name of parameter.}

\item{cov}{Character. Name of covariate.}

\item{...}{Additional arguments passed to \code{\link[dplyr:mutate]{dplyr::mutate()}}.}

\item{categorical}{Logical (default = \code{FALSE}).}

\item{plot_tv}{Logical.}
}
\value{
A \code{ggplot2} plot object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Plots posthoc parameter-covariate relationships from NONMEM run.
}
\details{
The mutate statement is to add variables not included in original
$TABLE.
}
