% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctl_handling.R
\name{param_info}
\alias{param_info}
\title{Get parameter information}
\usage{
param_info(ctl_name)
}
\arguments{
\item{ctl_name}{character. Path to control file}
}
\description{
Get parameter information
}
