% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_object.R
\name{parent_run}
\alias{parent_run}
\title{Get parent object of nm object}
\usage{
parent_run(m, n = 1L)
}
\arguments{
\item{m}{An nm object.}

\item{n}{Numeric. generation of parent (default = 1).}
}
\description{
Will pull the parent run of an nm object from the run cache.
}
