% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{perturb_inits}
\alias{perturb_inits}
\title{perturb initial estimates}
\usage{
perturb_inits(m, theta_log, omega_diag)
}
\arguments{
\item{m}{nm object}

\item{theta_log}{numeric (default = 0.1). Variance of theta perturbation}

\item{omega_diag}{numeric (default = 0.1). Variance of omega perturbation}
}
\description{
perturb initial estimates
}
