% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_iter.R
\name{plot_iter}
\alias{plot_iter}
\title{Plot iterations vs parameters/OBJ}
\usage{
plot_iter(nm_obj, trans = TRUE, ...)
}
\arguments{
\item{x}{object of class nm.execute}
}
\description{
Plot iterations vs parameters/OBJ
}

