% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_apps.R
\name{plot_iter_dygraph}
\alias{plot_iter_dygraph}
\title{Plot iterations vs parameters/OBJ with dygraphs}
\usage{
plot_iter_dygraph(m, trans = TRUE)
}
\arguments{
\item{m}{object of class nmexecute}

\item{trans}{logical. should parameter transformations be performed}
}
\description{
Plot iterations vs parameters/OBJ with dygraphs
}
