% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_iter.R
\name{plot_iter_ggplot}
\alias{plot_iter_ggplot}
\title{Plot iterations vs parameters/OBJ (ggplot2)}
\usage{
plot_iter_ggplot(d)
}
\arguments{
\item{d}{A \code{data.frame}. Output from \code{\link[=plot_iter_data]{plot_iter_data()}}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Used in shiny app and as a backend function.
}
\seealso{
\code{\link[=plot_iter_data]{plot_iter_data()}}
}
\keyword{internal}
