% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_iter.R
\name{plot_iter_plotly}
\alias{plot_iter_plotly}
\title{Plot iterations vs parameters/OBJ (plotly)}
\usage{
plot_iter_plotly(d)
}
\arguments{
\item{d}{data.frame. output from plot_iter_data}
}
\description{
Plot iterations vs parameters/OBJ (plotly)
}
