% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_reestExtra.R
\name{ppc_data}
\alias{ppc_data}
\title{process data from PPC for plotting}
\usage{
ppc_data(r, FUN, ..., max_mod_no = NA, DV = "DV", statistic = "statistic")
}
\arguments{
\item{r}{nm object}

\item{FUN}{statistic function with NONMEM dataset as arg and
returns data.frame with a column "statistic"}

\item{...}{additional arguments for FUN}

\item{DV}{character (default = "DV")}

\item{statistic}{character (default = "statistic") name of statistic column
returned by FUN}

\item{max_mod_mo}{integer. Maximum model number to read (set low for debugging)}
}
\description{
process data from PPC for plotting
}
