% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{preview}
\alias{preview}
\title{Preview code_library file}
\usage{
preview(name)
}
\arguments{
\item{name}{Character indicating script in code_library to preview.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

This is now primarily a back end function.  Easiest way to preview parts of
the code library is use the "code library" RStudio 'Addin' and use the
"preview" button in the shiny interface.
}
\seealso{
\code{\link[=code_library]{code_library()}}
}
\keyword{internal}
