% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_gettersetters.R
\name{prior_ctl}
\alias{prior_ctl}
\title{Add a prior ctl file contents to object}
\usage{
prior_ctl(m, ctl_ob, update_numbering = TRUE, update_dollar_data = TRUE,
  ...)
}
\arguments{
\item{m}{nm object}

\item{ctl_ob}{path to control file}

\item{update_numbering}{logical. Should table numbers and author fields be updated}

\item{update_dollar_data}{logical. Should $DATA in control file be updated}

\item{...}{additional arguments}
}
\description{
Add a prior ctl file contents to object
}
