% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMproject.R
\name{process_ppc}
\alias{process_ppc}
\title{Posterior predictive check computations}
\usage{
process_ppc(d, stat_fun, sim_col = "SIM", ...)
}
\arguments{
\item{d}{data.frame (normally output of nm_output())}

\item{stat_fun}{function to compute statistic.
First argument = data.frame for dataset
Function must use DV column}

\item{sim_col}{character (default = "SIM"). name of subproblem number column}

\item{...}{additional arguments to be passed to stat_fun}
}
\description{
Experimental function. Computes statistic on DV and DV_OUT for observed
and simulated datasets prospectively
}
\examples{
\dontrun{
do <- nm_output(m1)
cmax <- function(d){
  d \%>\% group_by(DOSE) \%>\% 
  summarise(Cmax = max(DV, na.rm = TRUE))
}
dppc <- process_ppc(d, cmax)
## then plot dppc$obs and dppc$sim
}
}
