% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-nonmem.R
\name{psn_available}
\alias{psn_available}
\title{Test if psn is available locally}
\usage{
psn_available()
}
\value{
Logical \code{TRUE} or \code{FALSE}
}
\description{
This is mainly for testing.  PsN doesn't have to be available locally for NMproject to work.  As long as it is available from \code{system_nm()}
}
\seealso{
\code{psn_check()}
}
\keyword{internal}
