% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-nonmem.R
\name{psn_check}
\alias{psn_check}
\title{Tests if psn is accessible}
\usage{
psn_check()
}
\value{
Logical TRUE or FALSE.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

NMproject needs PsN installed to function.  If PsN is not installed.
}
\examples{

psn_available()

}
\keyword{internal}
