% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psn_style_scm.R
\name{psn_style_scm}
\alias{psn_style_scm}
\title{PsN style stepwise covariate method}
\usage{
psn_style_scm(base, run_in, dtest, alpha_forward = 0.05, alpha_backward = 0.01)
}
\arguments{
\item{base}{nm object (base model)}

\item{run_in}{directory to run in}

\item{dtest}{output of test_relations}

\item{alpha_forward}{numeric (default = 0.05). alpha level for forward
inclusion}

\item{alpha_backward}{numeric (default = 0.01). alpha level for backward
deletion}
}
\description{
Intent is not to replicate PsN SCM.  This is mainly here for illustrative and
comparison purposes. Should replicate the model selection in PsN's SCM
functionality with greedy setting.
}
