% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{push_init_theta}
\alias{push_init_theta}
\alias{pull_init_theta}
\alias{push_init_theta}
\alias{pull_init_omega}
\alias{push_init_omega}
\alias{pull_init_sigma}
\alias{push_init_sigma}
\title{get/set parameter values}
\usage{
pull_init_theta(m)

push_init_theta(replace, m)

pull_init_omega(m)

push_init_omega(replace, m)

pull_init_sigma(m)

push_init_sigma(replace, m)
}
\arguments{
\item{m}{nm object}

\item{replace}{optional argument to replace}
}
\description{
get/set parameter values
}
\examples{
\dontrun{

m1 \%>\% pull_init_theta() \%>\%
  mutate_cond(
    name \%in\% "CL",
    init = 30
  ) \%>\%
  push_init_theta()

} 

}
