% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.table.nm.R
\name{read.table.nm}
\alias{read.table.nm}
\title{Fast read of NONMEM output table}
\usage{
read.table.nm(file, ...)
}
\arguments{
\item{file}{file argument - see ?read.table help for details}

\item{...}{other arguments to be passed to read.table()}
}
\description{
Fast read of NONMEM output table
}
