% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMproject.R
\name{read_derived_data}
\alias{read_derived_data}
\title{Read derived data}
\usage{
read_derived_data(name, ...)
}
\arguments{
\item{name}{name of file (without extension)}

\item{...}{additional arguments to be passed dto write.csv}
}
\description{
Read derived data
}
