% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R, R/run_record.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%<>\%}
\alias{\%T>\%}
\alias{\%$\%}
\alias{.data}
\alias{coef}
\alias{AIC}
\alias{nobs}
\alias{BIC}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr:exposition]{\%$\%}}, \code{\link[magrittr:compound]{\%<>\%}}, \code{\link[magrittr:tee]{\%T>\%}}}

  \item{rlang}{\code{\link[rlang:tidyeval-data]{.data}}}

  \item{stats}{\code{\link[stats]{AIC}}, \code{\link[stats:AIC]{BIC}}, \code{\link[stats]{coef}}, \code{\link[stats]{nobs}}}
}}

