% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{relative_path}
\alias{relative_path}
\title{compute path relative to reference}
\usage{
relative_path(path, relative_path)
}
\arguments{
\item{path}{character}

\item{relative_path}{character}
}
\description{
compute path relative to reference
}
