% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctl_handling.R
\name{remove_cov}
\alias{remove_cov}
\title{Remove a covariate to a NONMEM model}
\usage{
remove_cov(ctl, param, cov, state = 2, continuous = TRUE, time_varying,
  custom_state_text, id_var = "ID")
}
\arguments{
\item{ctl}{object coercible to ctl_list}

\item{param}{character. Name of parameter}

\item{cov}{character. Name of covariate}

\item{state}{numeric. Number of state}

\item{continuous}{logical (default = TRUE). is covariate continuous?}

\item{time_varying}{optional logical. is the covariate time varying?}

\item{custom_state_text}{optional character. custom state variable to be passed to param_cov_text}

\item{id_var}{character (default = "ID"). Needed if time_varying is missing.}
}
\description{
Only works with covariates added with add_cov
}
