% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param-manipulate.R
\name{rename_parameter}
\alias{rename_parameter}
\title{Rename a parameter in NONMEM control stream}
\usage{
rename_parameter(m, ...)
}
\arguments{
\item{m}{An nm object.}

\item{...}{Named arguments with character values indicated old names.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
\dontrun{

m1 \%>\% rename_parameter(CL = "KE")
## renames KE to CL

}
}
