% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMproject.R
\name{Renvironment_info}
\alias{Renvironment_info}
\title{Make R session record}
\usage{
Renvironment_info()
}
\description{
Create a record of the R version and package versions used in a particular NMproject
}
