% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{replace_code_library}
\alias{replace_code_library}
\title{Replaces code library}
\usage{
replace_code_library(path)
}
\arguments{
\item{path}{character vector with paths to attach to}
}
\description{
Replace code library search path with path(s)
}
