% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-run-summaries.R
\name{rr}
\alias{rr}
\title{Run record}
\usage{
rr(m, trans = TRUE)
}
\arguments{
\item{m}{An nm object.}

\item{trans}{Logical. If \code{TRUE} (default) will transform using control file
$THETA/OMEGA conventions.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Displays the transformed parameters of a completed or running model.
Normally used inside of a diagnostic template, but can be useful for quickly
seeing parameter estimates of several models.
}
\section{THETA transformations using \code{trans = TRUE}}{


Where FINAL is the reported estimate in the returned \code{tibble}, and RSE\% is
the reported standard error (where applicable) and \eqn{\theta} and \eqn{se(\theta)}
are the NONMEM reported values of parameters and standard errors, respectively

\describe{
\item{LOG or LOGODDS:}{
\eqn{FINAL = exp(\theta), RSE% = 100*\sqrt(exp^(se(\theta)) - 1)}
    }
\item{RATIO:}{
\eqn{FINAL = \theta, RSE% = 100*se(\theta) / \theta}
    }
\item{LOGIT:}{
\eqn{FINAL = 100 * 1 / (1 + exp(-\theta))}
}
}
}

\section{OMEGA transformations using \code{trans = TRUE}}{


Where FINAL is the reported estimate in the returned \code{tibble}, and RSE\% is
the reported standard error (where applicable) and \eqn{\omega^2} and \eqn{se(\omega^2)}
are the NONMEM reported values of parameters and standard errors, respectively

\describe{
\item{LOG:}{
\eqn{FINAL = 100*\sqrt(exp^(\omega^2) - 1), RSE% = 100*(se(\omega^2)/\omega^2)/2}
    }
}
}

\section{SIGMA transformations using \code{trans = TRUE}}{


Where FINAL is the reported estimate in the returned \code{tibble}, and RSE\% is
the reported standard error (where applicable) and \eqn{\sigma^2} and \eqn{se(\sigma^2)}
are the NONMEM reported values of parameters and standard errors, respectively

\describe{
\item{}{
\eqn{FINAL = \sqrt \omega^2, RSE% = 100*se(\omega^2) / \omega^2}
    }
}
}

\examples{
\dontrun{

rr(m1)

## compare m1 and m2

rr(c(m1, m2))
}
}
\seealso{
\code{\link[=nm_render]{nm_render()}}
}
