% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_record.R
\name{rr2}
\alias{rr2}
\title{run record}
\usage{
rr2(m, trans = TRUE)
}
\arguments{
\item{m}{nm object}

\item{trans}{logical. if TRUE (default) will transform using control file $THETA/OMEGA conventions}
}
\description{
run record
}
