% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{run_all_scripts}
\alias{run_all_scripts}
\title{Run all scripts sequentially}
\usage{
run_all_scripts()
}
\description{
Runs all scripts s01_..., s02_...
}
\details{
works with .R and .Rmd extensions.  Behaviour is to \link{source} .R
  files and use \link{rmarkdown::render} on .Rmd files
}
