% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_object.R
\name{run_dir_path}
\alias{run_dir_path}
\title{Get path to run_dir}
\usage{
run_dir_path(m)
}
\arguments{
\item{m}{An nm object.}
}
\description{
The function \code{\link[=run_dir]{run_dir()}} gives the directory name.  This function gets the
(relative) path of \code{\link[=run_dir]{run_dir()}}.
}
\seealso{
\code{\link[=nm_getsetters]{nm_getsetters()}}
}
