% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMproject.R
\name{run_id}
\alias{run_id}
\title{Get run id from file name or path}
\usage{
run_id(x)
}
\arguments{
\item{x}{character vector. name of file or path to get run id from}
}
\description{
Get run id from file name or path
}
