% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{run_in.nm_generic}
\alias{run_in.nm_generic}
\title{## run_in UseMethod already defined
#' @export
run_in.nm_generic <- function(x, text){
  if(missing(text)) {
    if(length(x$ctl_path) > 0) return(dirname(x$ctl_path)) else 
      return(NA_character_)
  }
  ctl_name <- basename(x$glue_fields["ctl_path"])
  x <- x %>% ctl_path(file.path(text, ctl_name))
  x
}
#' @export
run_in.nm_list <- Vectorize_nm_list(run_in.nm_generic)}
\usage{
\method{run_in}{nm_generic}(x, text)
}
\description{
## run_in UseMethod already defined
#' @export
run_in.nm_generic <- function(x, text){
  if(missing(text)) {
    if(length(x$ctl_path) > 0) return(dirname(x$ctl_path)) else 
      return(NA_character_)
  }
  ctl_name <- basename(x$glue_fields["ctl_path"])
  x <- x %>% ctl_path(file.path(text, ctl_name))
  x
}
#' @export
run_in.nm_list <- Vectorize_nm_list(run_in.nm_generic)
}
