% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{run_nm_fast_linux}
\alias{run_nm_fast_linux}
\title{experimental fast version of run (only tested in linux)}
\usage{
run_nm_fast_linux(r, wait = getOption("wait"), ignore.stdout = TRUE,
  ignore.stderr = TRUE, intern = getOption("intern"))
}
\arguments{
\item{r}{objects of class nm/or list of objects class nm/or data.frame with column of nm objects}

\item{wait}{logical (default=FALSE). Should R wait for run to finish.
Default can be changed with  wait_by_default() function}

\item{ignore.stdout}{logical (default=TRUE). Parameter passed to system()}

\item{ignore.stderr}{logical (default=TRUE). Parameter passed to system()}

\item{intern}{logical. intern arg to be passed to system}
}
\value{
If only one object of class nm was specified, silently returns object.
Otherwise returns nothing.
}
\description{
experimental fast version of run (only tested in linux)
}
