% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_record.R
\name{run_record}
\alias{run_record}
\title{Run record}
\usage{
run_record(..., trans = TRUE)
}
\arguments{
\item{...}{objects of class nmexecute}

\item{coef.func}{function to generate coefficients}
}
\description{
Run record
}

