% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_record.R
\name{run_record0}
\alias{run_record0}
\title{Run record}
\usage{
run_record0(..., coef.func = c(coef_ext0, coef.nm))
}
\arguments{
\item{...}{objects of class nmexecute}

\item{coef.func}{function to generate coefficients}
}
\description{
Run record
}

