% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{run_status}
\alias{run_status}
\title{run status}
\usage{
run_status(r, db, entry, initial_timeout = NA)
}
\arguments{
\item{r}{object class nm}

\item{db}{data.frame. Output of nmdb_get() (optional)}

\item{entry}{numeric. db entry number(optional)}

\item{initial_timeout}{numeric (default = NA).
Time to wait for directory creation before concluding error}
}
\description{
run status
}
