% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{set_simple_field}
\alias{set_simple_field}
\title{Add a simple field to nm object}
\usage{
set_simple_field(m, ...)
}
\arguments{
\item{m}{nm object}

\item{...}{arguments to set fields}
}
\description{
Add a simple field to nm object
}
\examples{
\dontrun{

core_list <- c(1,4,12)

mc <- m1 \%>\% child(run_id = paste0(corelist)) \%>\%
  set_simple_field(cores = corelist) \%>\%
  cmd("qpsn -c {cores} -t 59 -- execute {ctl_name} -dir={run_dir}")

}
}
