% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippet_setup.R
\name{setup_code_completion}
\alias{setup_code_completion}
\title{Set up code completion for NMproject}
\usage{
setup_code_completion(force = FALSE, snippet_path = find_snippet_path())
}
\arguments{
\item{force}{Logical.  The default is \code{FALSE} which will require user
confirmation before editing \code{r.snippets}.}

\item{snippet_path}{Character path to the \code{r.snippets} file.}
}
\value{
No return value, called for side effects.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Intelligent code completion is an experimental way to type NMproject code.
This function modifies/creates \code{r.snippets}.  Needs to be run interactively.
Will ask for user confirmation since snippets are an RStudio config setting
}
