% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippet_setup.R
\name{setup_nmproject_code_completion}
\alias{setup_nmproject_code_completion}
\title{Set up code completion for NMproject}
\usage{
setup_nmproject_code_completion(force = FALSE)
}
\arguments{
\item{force}{Logical.  The default is \code{FALSE} which will require user confirmation before editting \code{r.snippets}}
}
\value{
No return value, called for side effects.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Intelligent code completion is an experimental way to type NMproject code.
This function modifies/creates \code{r.snippets}.
}
