% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMproject.R
\docType{data}
\name{sge_parallel_execute}
\alias{sge_parallel_execute}
\title{Generic execute command for SGE grids}
\format{
An object of class \code{character} of length 1.
}
\usage{
sge_parallel_execute
}
\description{
Generic execute command for SGE grids
}
\details{
requires \code{cores} and \code{parafile} fields to be set
}
\seealso{
\code{\link{nm_getsetters}}
}
\keyword{datasets}
