% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{shiny_nm}
\alias{shiny_nm}
\title{Shiny view of NMproject}
\usage{
shiny_nm(m, envir = .GlobalEnv)
}
\arguments{
\item{m}{either nm_list object, or data.frame or list contain nm_lists}

\item{envir}{if missing, the environment to search}
}
\description{
Shiny view of NMproject
}
\examples{
\dontrun{

m1 \%>\% nm_shiny()
d$m \%>\% nm_shiny()
d \%>\% nm_shiny()

}
}
