% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{show_ctl}
\alias{show_ctl}
\title{Show an uneditable version of the control file}
\usage{
show_ctl(r)
}
\arguments{
\item{r}{nm object}
}
\description{
Show an uneditable version of the control file
}
