% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-file.R
\name{show_out}
\alias{show_out}
\title{Show an uneditable version of the lst file}
\usage{
show_out(r)
}
\arguments{
\item{r}{An nm object.}
}
\value{
No return value, called for side effects.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Opens a read-only version of the NONMEM control file for browsing.
}
\seealso{
\code{\link[=show_ctl]{show_ctl()}}.
}
