% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{show_runs}
\alias{show_runs}
\title{Show table of runs in database}
\usage{
show_runs(db_name = "runs.sqlite", ...)
}
\arguments{
\item{db_name}{character. Name of db}

\item{...}{arguments to pass to nmdb_get}
}
\description{
Show table of runs in database
}
