% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_object.R
\name{simple_field}
\alias{simple_field}
\title{Get/set simple fields of nm objects}
\usage{
simple_field(m, ...)
}
\arguments{
\item{m}{nm object}

\item{...}{arguments to get/set fields}
}
\description{
Get/set simple fields of nm objects
}
\examples{
\dontrun{

core_list <- c(1,4,12)

mc <- m1 \%>\% child(run_id = paste0(corelist)) \%>\%
  simple_field(cores = corelist) \%>\%
  cmd("qpsn -c {cores} -t 59 -- execute {ctl_name} -dir={run_dir}")
  
mc \%>\% simple_field(cores)  ## retrieve field from object

mc <- mc \%>\% simple_field(stars = 3)
mc \%>\% simple_field(stars)

}
}
