% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMproject.R
\name{snapshot}
\alias{snapshot}
\title{Git commit of ctl files, SourceData and Scripts}
\usage{
snapshot(message = "created automatic snapshot", session = TRUE,
  db_name = "runs.sqlite", ...)
}
\arguments{
\item{message}{character. Description to be added to commit}

\item{session}{logical. Should sessionInfo be included in commit message}

\item{db_name}{character. Name of database}

\item{...}{additional arguments for git2r::commit}
}
\description{
Git commit of ctl files, SourceData and Scripts
}
