% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{stage}
\alias{stage}
\title{stage files in project staging area ready for import}
\usage{
stage(
  files,
  destination,
  additional_sub_dirs = nm_dirs(),
  overwrite = FALSE,
  silent = FALSE
)
}
\arguments{
\item{files}{character vector. path of files to stage}

\item{destination}{default empty.  Optional destination directory
by default will be equivalent location in staging/}

\item{additional_sub_dirs}{character vector. additional subdirectories
not in standard tidyproject structure}

\item{overwrite}{logical (default = FALSE).}

\item{silent}{logical (default = FALSE)}
}
\description{
stage files in project staging area ready for import
}
