% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring.R
\name{status}
\alias{status}
\title{Get status of NONMEM runs}
\usage{
status(x)
}
\arguments{
\item{x}{An nm object.}
}
\value{
A character with the status of the run with values \code{"non started"},
\code{"running"}, \code{"finished"}, or \code{"error"}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\seealso{
\code{\link[=status_table]{status_table()}}.
}
