% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring.R
\name{status_table}
\alias{status_table}
\title{get status of multiple runs in form of table}
\usage{
status_table(m)
}
\arguments{
\item{m}{nm objects}
}
\description{
get status of multiple runs in form of table
}
