% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_record.R
\name{summary_long}
\alias{summary_long}
\title{Generate a long summary of NONMEM results}
\usage{
summary_long(..., parameters = c("none", "new", "all"))
}
\arguments{
\item{...}{arguments passed to summary(), usually a vector of nm object + options}

\item{parameters}{character. either \"none\" (default), \"new\", or \"all\"}
}
\description{
Generate a long summary of NONMEM results
}
