% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_record.R
\name{summary_wide}
\alias{summary_wide}
\title{Generate a wide summary of NONMEM results}
\usage{
summary_wide(
  ...,
  include_fields = character(),
  parameters = c("none", "new", "all"),
  m = TRUE,
  trans = TRUE
)
}
\arguments{
\item{...}{arguments passed to summary(), usually a vector of nm object + options}

\item{include_fields}{character}

\item{parameters}{character. either \"none\" (default), \"new\", or \"all\"}

\item{m}{logical (default = TRUE). should model object be included as m column}

\item{trans}{logical (default = TRUE). should parameters be transformed}
}
\description{
Generate a wide summary of NONMEM results
}
