% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{system_cmd}
\alias{system_cmd}
\title{System/shell command wrapper}
\usage{
system_cmd(cmd, dir = ".", ...)
}
\arguments{
\item{cmd}{Character. Command to send to shell.}

\item{dir}{Optional character. Directory to run command in (default = current
working directory)}

\item{...}{Other arguments passed to system command.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Will run \code{getOption("system_cmd")}.  A OS agnostic interface to the system
terminal.  Most of the time this will be the same as \code{system_nm} except when
the PsN/NONMEM execution server is location in a different location to the
RStudio server.
}
\examples{

\dontrun{

system_cmd("pwd")

}

}
