% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-nonmem.R
\name{system_nm_available}
\alias{system_nm_available}
\title{Test integrity of system_nm}
\usage{
system_nm_available(fail_if_false = FALSE)
}
\arguments{
\item{fail_if_false}{Logical (default = \code{FALSE}). Should the function return
an error message if \code{system_nm()} is not available.}
}
\value{
Logical \code{TRUE} or \code{FALSE}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{system_nm()} needs to be able to submit commands to the terminal for
NMproject to function.  This function tests whether meaningful output can be
produced and is only intended for diagnostic purposes.
}
\examples{

system_nm_available(fail_if_false = FALSE)

}
\keyword{internal}
