% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMproject.R
\name{system_nm_default}
\alias{system_nm_default}
\title{Default system_nm}
\usage{
system_nm_default(cmd, ...)
}
\arguments{
\item{cmd}{Character. system call.}

\item{...}{Additional args to be passed to system.}
}
\description{
Not intended to be used directly in most cases.
}
