% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-apps.R
\name{tail_lst}
\alias{tail_lst}
\title{Tail of lst file}
\usage{
tail_lst(r)
}
\arguments{
\item{r}{An nm object.}
}
\value{
A \code{character} vector of the tail end of the lst file.
}
\description{
Used by the shiny run monitor to show the tail end of the .lst file.
}
\keyword{internal}
