% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{target}
\alias{target}
\title{Target part of control object for further modification}
\usage{
target(m, dollar, lines)
}
\arguments{
\item{m}{nm object}

\item{dollar}{character. name of subroutine to target}

\item{lines}{optional character.  Assignment of lines}
}
\description{
Target part of control object for further modification
}
