% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_reestExtra.R
\name{test_relations}
\alias{test_relations}
\title{Generate tibble of relations to test}
\usage{
test_relations(dtest, param, cov, state, continuous)
}
\arguments{
\item{dtest}{(optional) existing dtest to append (from an previous use test_relations())}

\item{param}{character. Name of parameter}

\item{cov}{character. Name of covariate}

\item{state}{numeric or character. Number/name of state (see details)}

\item{continuous}{logical (default = TRUE). is covariate continuous?}
}
\description{
Generate tibble of relations to test
}
\details{
available states (see also \code{\link{add_cov}}):
"2" or "linear":
  PARCOV= ( 1 + THETA(1)*(COV - median))

"3" or "hockey-stick":
  IF(COV.LE.median) PARCOV = ( 1 + THETA(1)*(COV - median))
  IF(COV.GT.median) PARCOV = ( 1 + THETA(2)*(COV - median))
                    
"4" or "exponential":
   PARCOV= EXP(THETA(1)*(COV - median))

"5" or "power":
   PARCOV= ((COV/median)**THETA(1))
   
"power1":
   PARCOV= ((COV/median))
   
"power0.75":
   PARCOV= ((COV/median)**0.75)

"6" or "log-linear":
   PARCOV= ( 1 + THETA(1)*(LOG(COV) - log(median)))
}
\examples{

\dontrun{
dtest <- test_relations(param = c("KA", "K", "V"),
                        cov = c("LIN1", "LIN2", "LIN3", "RND1", "RND2", "RND3"), 
                        state = c("linear", "power"), 
                        continuous = TRUE)
dtest <- dtest \%>\% 
             test_relations(param = c("KA", "K", "V"),
                            cov = "BN1",
                            state = "linear",
                            continuous = FALSE)

}
}
\seealso{
\code{\link{add_cov}}, \code{\link{covariate_step_tibble}}
}
