% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctl_handling.R
\name{theta_nm2r}
\alias{theta_nm2r}
\title{Convert $THETA to R abstraction layer}
\usage{
theta_nm2r(x)
}
\arguments{
\item{x}{Character vector with $THETA information.}
}
\value{
Object abstracting $THETA contents.
}
\description{
Convert $THETA to R abstraction layer
}
\keyword{internal}
