% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctl_handling.R
\name{theta_r2nm}
\alias{theta_r2nm}
\title{Convert R abstraction layer to $THETA}
\usage{
theta_r2nm(x)
}
\arguments{
\item{x}{object abstracting $THETA contents}
}
\value{
character vector with $THETA information
}
\description{
Convert R abstraction layer to $THETA
}
