% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subroutine.R
\name{tol}
\alias{tol}
\title{Get/set TOL value in NONMEM control file}
\usage{
tol(m, text)
}
\arguments{
\item{m}{nm object}

\item{text}{optional character. value of TOL to set}
}
\description{
Get/set TOL value in NONMEM control file
}
