% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_object.R
\name{unblock}
\alias{unblock}
\title{Remove $OMEGA/$SIGMA BLOCK from init_omega and init_sigma output}
\usage{
unblock(iomega, eta_numbers)
}
\arguments{
\item{iomega}{tibble.  Output from init_omega() and init_sigma()}

\item{eta_numbers}{numeric vector.  ETA numbers to unblock. Must be contiguous}
}
\description{
Remove $OMEGA/$SIGMA BLOCK from init_omega and init_sigma output
}
\examples{

\dontrun{
io <- m1 \%>\% init_omega()
io <- io \%>\% unblock(c(2,3))
m1 <- m1 \%>\% init_omega(io)
m1 \%>\% dollar("OMEGA") ## to display $OMEGA
}

}
\seealso{
\code{\link[=block]{block()}}, \code{\link[=init_theta]{init_theta()}}
}
