% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctl_handling.R
\name{uncomment}
\alias{uncomment}
\title{Uncomment lines of control file}
\usage{
uncomment(m, pattern = ".*")
}
\arguments{
\item{m}{nm object}

\item{pattern}{optional character regex.  Passed to gsub}
}
\description{
Uncomment lines of control file
}
\seealso{
\code{\link{comment}}, @seealso \code{\link{gsub_ctl}}
}
