% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{untarget}
\alias{untarget}
\title{Remove target on control object}
\usage{
untarget(m, dollar)
}
\arguments{
\item{m}{nm object}

\item{dollar}{character. name of subroutine to target}
}
\description{
Remove target on control object
}
