% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMproject.R
\name{update_dollar_data}
\alias{update_dollar_data}
\title{update dollar data (name of dataset) in NONMEM control stream}
\usage{
update_dollar_data(ctl_name, new_data_name)
}
\arguments{
\item{ctl_name}{object coercible to ctl_character}

\item{new_data_name}{character. Name of new dataset}
}
\description{
Generic function
}
