% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMproject.R
\name{update_dollar_input}
\alias{update_dollar_input}
\title{update dollar inpute in NONMEM control stream}
\usage{
update_dollar_input(ctl, ...)
}
\arguments{
\item{ctl}{object coercible into ctl_lst}

\item{...}{arguments to be passed to dollar_input}
}
\description{
Generic function
}
