% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{update_ignore}
\alias{update_ignore}
\title{replace ignore statement}
\usage{
update_ignore(ctl, ignore_char)
}
\arguments{
\item{ctl}{object coercible into ctl_list}

\item{ignore_char}{character. replacement statement}
}
\description{
replace ignore statement
}
