% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctl_handling.R
\name{update_parameters}
\alias{update_parameters}
\title{update parameters from a control stream}
\usage{
update_parameters(ctl_lines, from)
}
\arguments{
\item{ctl_lines}{character vector. ctl file read into R}

\item{from}{class nm. object from which to extract results}
}
\description{
update parameters from a control stream
}
