% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{update_sizes}
\alias{update_sizes}
\title{update sizes statement}
\usage{
update_sizes(ctl, sizes_char)
}
\arguments{
\item{ctl}{object coercible into ctl_list}

\item{sizes_char}{character. replacement statement}
}
\description{
update sizes statement
}
