% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{valid_package_name}
\alias{valid_package_name}
\title{Package name validator from \code{usethis}}
\usage{
valid_package_name(x)
}
\arguments{
\item{x}{Name of package.}
}
\value{
A logical \code{TRUE} if or \code{FALSE} indicating if \code{x} is a valid package
name.
}
\description{
Package name validator from \code{usethis}
}
\keyword{internal}
