% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual-edit.R
\name{view_patch}
\alias{view_patch}
\title{View a patch}
\usage{
view_patch(patch_id)
}
\arguments{
\item{patch_id}{Character. Name of patch}
}
\value{
No return value, called for side effects.
}
\description{
Use the "view patch" RStudio 'Addin' for viewing patches instead.
}
\keyword{internal}
