% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{wait_default}
\alias{wait_default}
\title{Should run() wait for job to finish}
\usage{
wait_default(x)
}
\arguments{
\item{x}{logical. TRUE means run() will wait, FALSE = asynchronous execution}
}
\description{
Should run() wait for job to finish
}
