% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{wait_for_finished}
\alias{wait_for_finished}
\title{wait for a run to finish}
\usage{
wait_for_finished(..., timeout = NULL)
}
\arguments{
\item{...}{objects of class nm}

\item{timeout}{numeric. Maximum time (seconds) to wait}
}
\description{
wait for a run to finish
}

