% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_nm.R
\name{wipe_run}
\alias{wipe_run}
\title{Wipe previous run files}
\usage{
wipe_run(r)
}
\arguments{
\item{r}{An nm object.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Will remove all the output files generated by a previously completed run.
This is run by \code{\link[=run_nm]{run_nm()}} prior to launching any jobs to ensure that output
files from old runs do not get mistaken for up-to-date runs.
}
