% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{write_ctl}
\alias{write_ctl}
\title{Write control file to disk}
\usage{
write_ctl(m, prompt = TRUE)
}
\arguments{
\item{m}{nm object}

\item{prompt}{logical (default = TRUE), should prompt if overwriting?}
}
\description{
Normally used by other functions
}
