% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{write_derived_data_extra}
\alias{write_derived_data_extra}
\title{Write derived data file.}
\usage{
write_derived_data_extra(d, name, ...)
}
\arguments{
\item{d}{data.frame. Data frame to be saved}

\item{name}{name of file (without extension)}

\item{...}{additional arguments to be passed dto write.csv}
}
\description{
Write derived data file.
}
