% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{advan}
\alias{advan}
\alias{trans}
\alias{tol.nm_generic}
\title{get/set advan/trans/tol in ctl_contents}
\usage{
advan(m, text)

trans(m, text)

\method{tol}{nm_generic}(m, text)
}
\arguments{
\item{m}{nm object}

\item{text}{optional number/character number to set to}
}
\description{
Will make no other changes to control file
}
