% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_edit.R
\name{apply_manual_edit}
\alias{apply_manual_edit}
\title{apply a manual edit}
\usage{
apply_manual_edit(m, patch_name)
}
\arguments{
\item{m}{nm object}

\item{patch_name}{character name of patch}
}
\description{
Best used with "make manual edit patch" addin.
}
