% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_manual_edit.R
\name{apply_manual_edit}
\alias{apply_manual_edit}
\title{Apply a manual edit patch}
\usage{
apply_manual_edit(m, patch_name)
}
\arguments{
\item{m}{nm object}

\item{patch_name}{character name of patch}
}
\description{
Best to allow the "manual edit" RStudio addin to write this function for you
}
\details{
Generally best to to apply patches before automatic edits and
  changes in directories e.g. via \code{run_in()}
}
