% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_methods.R
\name{as_nm_list}
\alias{as_nm_list}
\title{Coerce object into nm_list}
\usage{
as_nm_list(m)
}
\arguments{
\item{m}{An object coercible into an nm object.}
}
\description{
This may be relegated to a back end function soon.  Create an nm_list from an
nm_generic or list or other object coercible into an nm_list.
}
\keyword{internal}
