% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{attach_code_library}
\alias{attach_code_library}
\title{Attach code library}
\usage{
attach_code_library(path)
}
\arguments{
\item{path}{character vector with paths to attach to}
}
\description{
Attaches a path(s) to to the code library search path
}
