% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_reestExtra.R
\name{bind_covariate_results}
\alias{bind_covariate_results}
\title{Add run results into a covariate tibble}
\usage{
bind_covariate_results(dsc, nm_col = "m", parameters = "new")
}
\arguments{
\item{dsc}{a covariate tibble (see \code{\link[=covariate_step_tibble]{covariate_step_tibble()}})}

\item{nm_col}{character (default = "m"). name of column to store nm objects}

\item{parameters}{character (default = "new").  Passed to \code{\link[=summary_wide]{summary_wide()}}}
}
\description{
Add run results into a covariate tibble
}
\examples{
\dontrun{
## create tibble of covariate step with model objects as column m
dsm1 <- m1 \%>\% covariate_step_tibble(
  run_id = "m1_f1",
  dtest = dtest,
  direction = "forward"
)

## run all models greedily
dsm1$m <- dsm1$m \%>\% run_nm()

## extract results and put into tibble
dsm1 <- dsm1 \%>\% bind_covariate_results()
}
}
\seealso{
\code{\link[=covariate_step_tibble]{covariate_step_tibble()}}
}
