% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_object.R
\name{change_parent}
\alias{change_parent}
\title{Change parent object}
\usage{
change_parent(m, mparent, silent = TRUE)
}
\arguments{
\item{m}{nm object}

\item{mparent}{new parent object to child}

\item{silent}{logical (default = TRUE). Should warn if conflicts detected}
}
\description{
When a \code{child()} has been created by a modified parent sometimes adoption is
the answer.  This will force parenthood to the specified object.
}
\seealso{
\code{\link[=child]{child()}}
}
\keyword{internal}
