% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_object.R
\name{change_parent}
\alias{change_parent}
\title{Change parent object}
\usage{
change_parent(m, mparent, silent = TRUE)
}
\arguments{
\item{m}{An nm object.}

\item{mparent}{New parent object to child.}

\item{silent}{Logical (default = \code{TRUE}). Should warn if conflicts detected.}
}
\value{
An nm object with modified \verb{parent_*} fields.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

When a \code{child()} has been created by a modified parent sometimes adoption is
the answer.  This will force parenthood to the specified object.
}
\seealso{
\code{\link[=child]{child()}}
}
\keyword{internal}
