% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_installation.R
\name{check_installation}
\alias{check_installation}
\title{Check NMproject installation}
\usage{
check_installation()
}
\value{
Logical \code{TRUE} or \code{FALSE} indicating whether the tests have succeeded
or not.  A test failure does not necessarily mean that NMproject
incorrectly configured.  Test messages will say whether they are needed or
just recommended to pass.
}
\description{
Check NMproject installation
}
