% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_object.R
\name{child}
\alias{child}
\title{Make child nm object from parent}
\usage{
child(m, run_id = NA_character_, type = "execute", silent = FALSE)
}
\arguments{
\item{m}{parent nm object}

\item{run_id}{character.  New run id to assign to child object}

\item{type}{character (default = "execute"). type of child object}

\item{silent}{logical (default = FALSE). Should warn if conflicts detected}
}
\description{
Child objects inherit attributes of parent but with a new run_id.
The control file will be inherited too with $TABLEs updated
}
\examples{
\dontrun{

m2 <- m1 \%>\% child("m2")

}
}
