% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{clean_run}
\alias{clean_run}
\title{clean_run files}
\usage{
clean_run(r, delete_dir = c(NA, TRUE, FALSE),
  update_db = !is.null(r$db_name))
}
\arguments{
\item{r}{object class nm}

\item{delete_dir}{logical. NA (default), TRUE or FALSE. Should run_dir be deleted.}

\item{update_db}{logical (default=TRUE). Should run_status be updated}
}
\description{
clean_run files
}
