% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{code_library}
\alias{code_library}
\title{Code Library}
\usage{
code_library(
  extn = NULL,
  fields = "Description",
  viewer = TRUE,
  silent = FALSE,
  return_info = FALSE
)
}
\arguments{
\item{extn}{vector string of extensions to include}

\item{fields}{character vector of fields to extract}

\item{viewer}{logical indicating if viewer should be used to display results
(default=FALSE)}

\item{silent}{logical indicating if messages should be silenced
(default=FALSE)}

\item{return_info}{logical (default = FALSE). Return data.frame of results
(FALSE= returns file paths)}
}
\description{
Function not designed for direct use.  Instead use the RStudio "addin" on the
Addins menu. In the shiny, select the file, and click "preview" to view and
\code{\link{stage}} to bring into the "staging" area of your project.  See
vignette at \url{http://tsahota.github.io/NMproject} for a video showing use
of the app. Non-NONMEM code can then be imported into the project separately
with the the \code{\link{import}} function.
}
\details{
requires \code{getOption("code_library_path")} to be set
}
\seealso{
\code{\link{ls_code_library}}, \code{\link{preview}},
  \code{\link{stage}}, \code{\link{import}}
}
