% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{coef_long}
\alias{coef_long}
\title{coef (in long format)}
\usage{
coef_long(m, trans = TRUE)
}
\arguments{
\item{m}{nm object}

\item{trans}{logical (default = TRUE) should transform parameters?}
}
\description{
coef (in long format)
}
