% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_record.R
\name{coef_nm}
\alias{coef_nm}
\title{Gets parameters values}
\usage{
coef_nm(object, trans, ...)
}
\arguments{
\item{object}{object of class nmexecute}

\item{trans}{logical. Default = TRUE. Should parameters be transformed}

\item{...}{additional arguments to carry to coef}
}
\value{
a \code{data.frame} of parameter values
}
\description{
Produces data.frame of parameter values
}
