% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-run-summaries.R
\name{coef_widelong}
\alias{coef_widelong}
\alias{coef_wide}
\alias{coef_long}
\title{Extract parameter values}
\usage{
coef_wide(m, trans = TRUE)

coef_long(m, trans = TRUE)
}
\arguments{
\item{m}{An nm object.}

\item{trans}{Logical (default = \code{TRUE}). Transform parameters using comments
in $THETA/$OMEGA/$SIGMA.}
}
\value{
\code{data.frame} of extracted model parameter values. \code{coef_wide()}
returns a \code{data.frame} in wide format. Vector valued objects \code{m}, will be
stacked vertically with one row per run.  \code{coef_long()} returns a
\code{data.frame} in long format. Vector valued objects \code{m}, will be stacked
horizontally.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Pulls parameters, standard errors, OFVs and condition numbers out of ext
files, applies transformations.  This function is useful when numeric values
are needed. \code{rr} is easier to read, however it returns characters.  A wide
and long format is available via two different functions.
}
\seealso{
\code{\link[=rr]{rr()}}
}
