% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctl_handling.R
\name{comment_out}
\alias{comment_out}
\title{Comment lines of control file}
\usage{
comment_out(m, pattern = ".*")
}
\arguments{
\item{m}{nm object}

\item{pattern}{optional character regex.  Passed to gsub}
}
\description{
Comment lines of control file
}
\seealso{
\code{\link[=uncomment]{uncomment()}}, @seealso \code{\link[=gsub_ctl]{gsub_ctl()}}
}
