% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_record.R
\name{cond_num}
\alias{cond_num}
\title{Condition number of run}
\usage{
cond_num(r)
}
\arguments{
\item{r}{An nm object}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Extracts condition number from .ext file.
}
\examples{

\dontrun{

is_finished(m1)  ## should be TRUE
cond_num(m1) ## displays condition number

}

}
\seealso{
\code{\link[=ofv]{ofv()}}, \code{\link[=rr]{rr()}}
}
