% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_object.R
\name{convert_to_simulation}
\alias{convert_to_simulation}
\title{convert a NONMEM run to a simulation}
\usage{
convert_to_simulation(m, seed = 12345, subpr = 1)
}
\arguments{
\item{m}{nm object}

\item{seed}{numeric (default = 12345). seed value to include in $SIM}

\item{subpr}{numeric (default = 1). SUBPR value to include in $SIM}
}
\description{
replaces $EST with $SIM
}
\details{
will only change $EST/$SIM, therefore it will not be sufficient to change a categorical estimation control file to simulation
}
