% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{cov_forest_data}
\alias{cov_forest_data}
\title{produce dataset for covariate forest plotting}
\usage{
cov_forest_data(m, id_col = "ID")
}
\arguments{
\item{m}{nm object}

\item{id_col}{character (default = "ID"). name of column}
}
\description{
produce dataset for covariate forest plotting
}
