% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_building.R
\name{cov_forest_data}
\alias{cov_forest_data}
\title{produce dataset for covariate forest plotting}
\usage{
cov_forest_data(m, covariate_scenarios)
}
\arguments{
\item{m}{nm object}

\item{covariate_scenarios}{data.frame. need names "cov", "value" and (optional) "text"}
}
\description{
produce dataset for covariate forest plotting
}
\examples{
\dontrun{

dcov <- input_data(m1, filter = TRUE)
dcov <- dcov[!duplicated(dcov$ID), ]
covariate_scenarios <- bind_rows(
  tibble(cov = "HEALTHGP", value = c(0, 1)),
  tibble(cov = "HEPATIC", value = unique(dcov$HEPATIC[dcov$HEPATIC > -99])),
  tibble(cov = "BWTIMP", value = c(50, 80, 120)),
  tibble(cov = "ECOG", value = c(0,1,2,3)),
  tibble(cov = "BEGFRIMP", value = quantile(dcov$BEGFR[dcov$BEGFR > -99])),
  tibble(cov = "RACE", value = c(1,2),text=c("white","black")),
  tibble(cov = "PPI", value = c(0,1)),
  tibble(cov = "H2RA", value = c(0,1))
)

dplot <- cov_forest_data(m1, covariate_scenarios = covariate_scenarios)
cov_forest_plot(dplot)

}

}
