% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{cov_forest_plot}
\alias{cov_forest_plot}
\title{plotting covariate forest plots}
\usage{
cov_forest_plot(d)
}
\arguments{
\item{d}{data.frame from cov_forest_data}
}
\description{
plotting covariate forest plots
}
