% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-run-summaries.R
\name{covariance_matrix}
\alias{covariance_matrix}
\title{Get covariance matrix}
\usage{
covariance_matrix(m)
}
\arguments{
\item{m}{An nm object.}
}
\value{
A \code{matrix} with parameter correlations.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Extracts covariance matrix
}
\examples{

## requires NONMEM to be installed
\dontrun{

covariance_matrix(m)

}
}
