% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_record.R
\name{covariance_plot}
\alias{covariance_plot}
\title{Plot covariance matrix}
\usage{
covariance_plot(r, trans = TRUE)
}
\arguments{
\item{r}{nm object}

\item{trans}{logical (default = TRUE)}
}
\description{
Plot covariance matrix
}
