% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_reestExtra.R
\name{covariate_step_tibble}
\alias{covariate_step_tibble}
\title{Prepare forward covariate step}
\usage{
covariate_step_tibble(
  base,
  run_id,
  run_in = nm_default_dir("models"),
  dtest,
  direction = c("forward", "backward"),
  ...
)
}
\arguments{
\item{base}{nm object}

\item{run_id}{base run_id to construct run_ids of covariate runs}

\item{run_in}{character.  See \code{\link{run_in}}}

\item{dtest}{tibble::tibble with testing relations (from \code{\link{test_relations}})}

\item{direction}{character. "forward" (default) or "backward"}

\item{...}{additional arguments passed to add_cov}
}
\value{
tibble::tibble dtest with appended columns
}
\description{
Prepare forward covariate step
}
\examples{
\dontrun{

dtest <- test_relations(param = c("KA", "K", "V"),
                        cov = c("LIN1", "LIN2", "LIN3", "RND1", "RND2", "RND3"), 
                        state = c("linear", "power"), 
                        continuous = TRUE)
dtest <- dtest \%>\% 
             test_relations(param = c("KA", "K", "V"),
                            cov = "BN1",
                            state = "linear",
                            continuous = FALSE)
                            
## create tibble of covariate step with model objects as column m
dsm1 <- m1 \%>\% covariate_step_tibble(
  run_id = "m1_f1",
  dtest = dtest,
  direction = "forward"
)

## run all models greedily
dsm1$m <- dsm1$m \%>\% run_nm()

## extract results and put into tibble
dsm1 <- dsm1 \%>\% bind_covariate_results()

## sort by BIC (for example) and view
dsm1 <- dsm1 \%>\% arrange(BIC)
dsm1

## check condition number, covariance,... 
## run any diagnostics here

## when happy with selection, select run for subsequent step

m1_f1 <- dsm1$m[1]   ## select most signifcant BIC
# alternative select by relationship
m1_f1 <- dsm1 \%>\%
    filter(param = "CL", cov = "BWT", state = "power") \%$\%
    m

## do next forward step

dsm2 <- m1_f1 \%>\% covariate_step_tibble(
  run_id = "m1_f2",
  dtest = dtest,
  direction = "forward"
)

## continue ...
  
}
}
\seealso{
\code{\link{test_relations}}, \code{\link{covariate_step_tibble}}, \code{\link{bind_covariate_results}}
}
