% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low-level-ctl-handling-funs.R
\name{ctl_character}
\alias{ctl_character}
\title{Constructor/converter to ctl_character}
\usage{
ctl_character(r)
}
\arguments{
\item{r}{Either class \code{nmexecute}, \code{character}, \code{ctl_list}, \code{ctl_character}.}
}
\value{
Object of class \code{ctl_character}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Convert a variety of objects into a \code{ctl_character} class
}
\keyword{internal}
