% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-gettersetters.R
\name{ctl_contents}
\alias{ctl_contents}
\title{Get/set control file contents}
\usage{
ctl_contents(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{\link[=based_on]{based_on()}}.}
}
\value{
An nm object with modified \code{ctl_contents} field.
}
\description{
This function is an alias for \code{\link[=based_on]{based_on()}}.
}
\examples{

# create example object m1 from package demo files
exdir <- system.file("extdata", "examples", "theopp", package = "NMproject")
m1 <- new_nm(run_id = "m1", 
             based_on = file.path(exdir, "Models", "ADVAN2.mod"),
             data_path = file.path(exdir, "SourceData", "THEOPP.csv"))
             
m1 \%>\% ctl_contents()

}
