% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-gettersetters.R
\name{ctl_contents}
\alias{ctl_contents}
\title{Get/set control file contents}
\usage{
ctl_contents(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{\link[=based_on]{based_on()}}.}
}
\description{
This function is an alias for \code{\link[=based_on]{based_on()}}.
}
