% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctl_handling.R
\name{ctl_list}
\alias{ctl_list}
\title{Constructor/converter to ctl_list}
\usage{
ctl_list(r)
}
\arguments{
\item{r}{either class nmexecute, character, ctl_list, ctl_character}
}
\value{
object of class ctl_list
}
\description{
Internal function
}
