% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low-level-ctl-handling-funs.R
\name{ctl_list}
\alias{ctl_list}
\title{Constructor/converter to ctl_list}
\usage{
ctl_list(r)
}
\arguments{
\item{r}{Either class \code{nmexecute}, \code{character}, \code{ctl_list}, \code{ctl_character}.}
}
\value{
Object of class \code{ctl_list}.
}
\description{
Constructor/converter to ctl_list
}
\keyword{internal}
