% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{ctl_path}
\alias{ctl_path}
\title{Path to NONMEM control file}
\usage{
ctl_path(m, text)
}
\arguments{
\item{m}{nm object}

\item{text}{optional character. Name of path to control file (see details).
Typically, this file does not yet normally exist, but will house the code
code for this run}
}
\description{
Get and set path to control file.
}
\details{
"text" can contain "{run_id}" string.  E.g. "Models/run{run_id}.mod"
  will use the name "Models/runm1.mod" if run_id(m1) is "m1".
}
\examples{
\dontrun{

m0 <- nm2(run_id = "m0")
ctl_path(m0)
m0 <- m0 \%>\% ctl_path("Models/run_m0.mod")
m0
## warning: this has make the run identifier in old xpose4 "_m0" not "m0"!

m1 <- m0 \%>\% run_id("m1")
ctl_path(m1)  ## run_id number has updated but ctl_path hasn't changed

## To avoid this use {run_id}
m0 <- m0 \%>\% ctl_path("Models/run_{run_id}.mod")
ctl_path(m0)

m1 <- m0 \%>\% run_id("m1")
m1  ## ctl_path has updated as expected

}
}
