% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_gettersetters.R
\name{ctl_path}
\alias{ctl_path}
\title{Path to NONMEM control file}
\usage{
ctl_path(m, text)
}
\arguments{
\item{m}{An nm object.}

\item{text}{Optional character. Name of path to control file (see details).
Typically, this file does not yet normally exist, but will house the code
code for this run.}
}
\description{
Get and set path to control file.
}
\details{
"text" can contain "{run_id}" string.  E.g. "Models/run{run_id}.mod"
  will use the name "Models/runm1.mod" if run_id(m1) is "m1".
}
\examples{
\dontrun{

m0 <- nm(run_id = "m0")
ctl_name(m0)

m0 <- m0 \%>\% ctl_name("run{run_id}.mod")
ctl_name(m0)

## warning: this makes the run identifier in old xpose4 "_m0" not "m0"!
m0 <- m0 \%>\% ctl_path("Models/run_{run_id}.mod")
ctl_path(m0)


}
}
