% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filter.R
\name{data_filter_char}
\alias{data_filter_char}
\title{Get filter statement}
\usage{
data_filter_char(r, ...)
}
\arguments{
\item{r}{Object coercible into ctl_list.}

\item{...}{Arguments passed to data_ignore_char.}
}
\value{
Negation of character value of \code{\link[=data_ignore_char]{data_ignore_char()}}.
}
\description{
Opposite of \code{\link[=data_ignore_char]{data_ignore_char()}}.
}
\keyword{internal}
