% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{decision}
\alias{decision}
\title{make decision point}
\usage{
decision(inputs = c(), files = c(), auto_decision = TRUE, outcome)
}
\arguments{
\item{inputs}{(optional) non file names upon which decision depends}

\item{files}{(optional) file names upon which decision depends}

\item{auto_decision}{(optional) logical. logical statement for automatic decisions}

\item{outcome}{character. Description of the decision outcome}
}
\description{
formalise process of decision making.  Creates a decision point.
Requests inputs (\code{values} and \code{files}) that you base a
decision on and stop for users to remake decision if inputs change
}
