% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{delete_dollar}
\alias{delete_dollar}
\title{Deleted subroutine}
\usage{
delete_dollar(m, dollar)
}
\arguments{
\item{m}{nm object}

\item{dollar}{character. Name of subroutine}
}
\description{
Deleted subroutine
}
