% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_object.R
\name{delete_dollar}
\alias{delete_dollar}
\title{Delete a NONMEM subroutine from control file contents}
\usage{
delete_dollar(m, dollar)
}
\arguments{
\item{m}{An nm object.}

\item{dollar}{Character. Name of subroutine.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
