% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{dollar}
\alias{dollar}
\title{Get/set existing subroutine}
\usage{
dollar(m, dollar, ...)
}
\arguments{
\item{m}{nm object}

\item{dollar}{character.  Subroutine to target}

\item{...}{arguments to be passed to text()}
}
\description{
Get/set existing subroutine
}
