% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic-ctl-manipulation.R
\name{dollar}
\alias{dollar}
\title{Get/set existing subroutine}
\usage{
dollar(m, dollar, ..., add_dollar_text = TRUE)
}
\arguments{
\item{m}{An nm object.}

\item{dollar}{Character. Name of NONMEM subroutine to target.}

\item{...}{Additional arguments to be passed to \code{text()}.  If specified these
will set the contents of the subroutine.  See examples below.}

\item{add_dollar_text}{Logical (default = \code{TRUE}). Should the $XXXX string be
added to text.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

The fast way to see the contents of a particular subroutine directly in the R
console. It can also be used to set the contents of a NONMEM subroutine in
place of manual edits
}
\examples{
\dontrun{

m1 \%>\% dollar("PK") ## displays existing $PK

m1 <- m1 \%>\% dollar(
  "DES",
  "
DADT(1) = -K*A(1)
"
)

## This will rewrite an existing $DES
## if control file doesn't already have a $DES
## use insert_dollar() instead
}

}
\seealso{
\code{\link[=insert_dollar]{insert_dollar()}}, \code{\link[=delete_dollar]{delete_dollar()}}
}
