% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{fill_input}
\alias{fill_input}
\title{set $INPUT}
\usage{
fill_input(m, ...)
}
\arguments{
\item{m}{nm object}

\item{keep}{character vector. Names of columns to keep}

\item{rename}{named character vector. Renaming instructions}
}
\description{
set $INPUT
}
\examples{
\dontrun{

 m1 <- m1 \%>\% fill_input(rename = c("DAT0" = "DATE"))

}
}
