% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{fill_input}
\alias{fill_input}
\title{Get/set $INPUT}
\usage{
fill_input(m, ...)
}
\arguments{
\item{m}{nm object}

\item{...}{optional additional arguments passed to dollar_data()}
}
\description{
Get/set $INPUT
}
