% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMproject.R
\name{from_models}
\alias{from_models}
\title{Path of directory from models dir}
\usage{
from_models(x, models_dir = nm_default_dir("models"))
}
\arguments{
\item{x}{character vector. Relative path from models.dir}

\item{models_dir}{character. Models directory}
}
\description{
Path of directory from models dir
}
