% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMproject.R
\name{from_models}
\alias{from_models}
\title{path of directory from models dir}
\usage{
from_models(x)
}
\arguments{
\item{x}{character vector. relative path from models.dir}
}
\description{
path of directory from models dir
}

