% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise-covariate.R
\name{gen_sim_path}
\alias{gen_sim_path}
\title{Generate paths for simulation runs}
\usage{
gen_sim_path(dsc, start_dir, include_names = TRUE)
}
\arguments{
\item{dsc}{A \code{tibble} with experimental factors to vary.}

\item{start_dir}{Directory name within which all runs should take place.}

\item{include_names}{Logical (default = TRUE). Should names be included.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Internal function to create nested directory structure paths.
}
\examples{
\dontrun{

dsc$location <- gen_sim_path(dsc, ...)

}
}
\keyword{internal}
