% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMproject.R
\name{get_PMX_code_library}
\alias{get_PMX_code_library}
\title{Download code repositor from github}
\usage{
get_PMX_code_library(local_path,
  giturl = "https://github.com/tsahota/PMXcodelibrary", config_file)
}
\arguments{
\item{local_path}{character. Path to install repository}

\item{giturl}{character. URL to github repository}

\item{config_file}{character. Name of config file (e.g. "~/.Rprofile")}
}
\description{
Download code repositor from github
}
