% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctl_handling.R
\name{gsub_ctl}
\alias{gsub_ctl}
\title{gsub for ctl file}
\usage{
gsub_ctl(ctl, ...)
}
\arguments{
\item{ctl}{object coercible into ctl_character}

\item{...}{arguments passed to gsub}
}
\description{
gsub for ctl file
}
