% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctl_handling.R
\name{gsub_ctl}
\alias{gsub_ctl}
\title{gsub for ctl file}
\usage{
gsub_ctl(ctl, pattern, replacement, ..., dollar = NA_character_)
}
\arguments{
\item{ctl}{object coercible into ctl_character}

\item{pattern}{argument passed to gsub}

\item{replacement}{argument passed to gsub}

\item{dollar}{character name of subroutine}
}
\description{
gsub for ctl file
}
