% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{ignore}
\alias{ignore}
\title{Get/set ignore statement from ctl_contents}
\usage{
ignore(ctl, ignore_char)
}
\arguments{
\item{ctl}{nm object}

\item{ignore_char}{optional character. Ignore statement}
}
\description{
Get/set ignore statement from ctl_contents
}
