% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{info_scripts}
\alias{info_scripts}
\title{List information about scripts}
\usage{
info_scripts(
  files,
  fields = c("Description"),
  viewer = TRUE,
  silent = FALSE,
  base_dirs = NULL,
  shorten_paths = TRUE
)
}
\arguments{
\item{files}{Vector string of file names/paths.}

\item{fields}{Vector string of field tags to display.}

\item{viewer}{Logical indicating if Rstudio viewer should be used (default =
\code{TRUE}).}

\item{silent}{Run in quiet mode (default=\code{FALSE}).}

\item{base_dirs}{Character vector. group files together that belong to these
directory paths.}

\item{shorten_paths}{Logical (default = \code{TRUE}). Long paths will be shortened
if true in displayed output (not returned object).}
}
\description{
List information about scripts
}
\examples{
\dontrun{
ls_scripts('~/path/to/directory/') \%>\%
  info_scripts('Description') \%>\%
  filter(grepl('mod',DESCRIPTION))
}
}
\keyword{internal}
