% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{init_theta}
\alias{init_theta}
\alias{init_omega}
\alias{init_sigma}
\title{Get/set initial parameters}
\usage{
init_theta(m, replace, ...)

init_omega(m, replace, ...)

init_sigma(m, replace, ...)
}
\arguments{
\item{m}{nm object}

\item{replace}{optional tibble for replacement}

\item{...}{mutate init_theta}
}
\description{
Get/set initial parameters
}
\examples{
\dontrun{

 ## fix THETA on KA
  ip <- init_theta(m1)
  ip$FIX[ip$name \%in\% "KA"] <- TRUE
  m1 <- m1 \%>\% init_theta(ip)
  
 ## fix OMEGA on KA to 0.0225
  ip <- init_omega(m1)
  ip$init[ip$name \%in\% "IIV_KA"] <- 0.0225
  ip$FIX[ip$name \%in\% "IIV_KA"] <- TRUE
  m1 <- m1 \%>\% init_omega(ip)
  
 ## perturb all log transformed parameters by ~10\%
  ip <- init_theta(m1)
  ip <- ip \%>\% mutate(
    init = ifelse(
      trans \%in\% "LOG",
      rnorm(length(init), mean = init, sd = 0.1),
      init
    )
  )
  m1 <- m1 \%>\% init_theta(ip)
 

}
}
