% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{insert_dollar}
\alias{insert_dollar}
\title{Insert a new subroutine into ctl_contents}
\usage{
insert_dollar(m, dollar, text, after_dollar)
}
\arguments{
\item{m}{nm object}

\item{dollar}{character. Name of subroutine}

\item{text}{character vector. Text to fill}

\item{after_dollar}{character. Positioning of subroutine}
}
\description{
Insert a new subroutine into ctl_contents
}
