% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring.R
\name{is_finished}
\alias{is_finished}
\title{Tests if job is finished}
\usage{
is_finished(r, initial_timeout = NA)
}
\arguments{
\item{r}{An nm object.}

\item{initial_timeout}{Deprecated.  See \code{\link[=wait_finish]{wait_finish()}}.}
}
\value{
A logical vector with \code{TRUE} or \code{FALSE} values.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{

# create example object m1 from package demo files
exdir <- system.file("extdata", "examples", "theopp", package = "NMproject")
m1 <- new_nm(run_id = "m1", 
             based_on = file.path(exdir, "Models", "ADVAN2.mod"),
             data_path = file.path(exdir, "SourceData", "THEOPP.csv"))

is_finished(m1) # FALSE

}
