% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_methods.R
\name{is_nm}
\alias{is_nm}
\alias{is_nm_list}
\alias{is_nm_generic}
\title{Test if object is an nm coercible object}
\usage{
is_nm_list(x)

is_nm_generic(x)
}
\arguments{
\item{x}{Object.}
}
\value{
A logical vector with \code{TRUE} or \code{FALSE} values.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Mostly internal functions to test object types.
}
