% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring.R
\name{is_successful}
\alias{is_successful}
\title{Test if NONMEM ran without errors}
\usage{
is_successful(r)
}
\arguments{
\item{r}{An nm object.}
}
\value{
\code{TRUE} if run was successful, \code{FALSE} otherwise.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{

# create example object m1 from package demo files
exdir <- system.file("extdata", "examples", "theopp", package = "NMproject")
m1 <- new_nm(run_id = "m1", 
             based_on = file.path(exdir, "Models", "ADVAN2.mod"),
             data_path = file.path(exdir, "SourceData", "THEOPP.csv"))
             
is_successful(m1) ## FALSE

}
