% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{job_info}
\alias{job_info}
\title{Get job information (if it exists)}
\usage{
job_info(m, text)
}
\arguments{
\item{m}{object class nm}

\item{text}{optional character to set job_info}
}
\description{
Requires 'get_job_info' to be defined as an option - function that takes stdout console output
from a job and returns a character
}
