% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system-hooks.R
\name{job_info}
\alias{job_info}
\title{Get job information (if it exists)}
\usage{
job_info(m, text)
}
\arguments{
\item{m}{An nm object.}

\item{text}{Optional character to set job_info.}
}
\value{
The value of the \code{job_info} field of \code{m} if \code{text} is missing.
Otherwise an nm object with modified \code{job_info} field.
}
\description{
Can be useful for storing job numbers with a modified \code{\link[=system_nm]{system_nm()}}.
}
\seealso{
\code{\link[=system_nm]{system_nm()}}
}
\keyword{internal}
