% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_info.R
\name{job_info}
\alias{job_info}
\title{Get job information (if it exists)}
\usage{
job_info(m, text)
}
\arguments{
\item{m}{An nm object.}

\item{text}{Optional character to set job_info.}
}
\description{
Can be useful for storing job numbers with a modified \code{\link[=system_nm]{system_nm()}}.
}
\seealso{
\code{\link[=system_nm]{system_nm()}}
}
\keyword{internal}
