% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_stats.R
\name{job_stats}
\alias{job_stats}
\title{Get job stats for a completed NONMEM run}
\usage{
job_stats(m)
}
\arguments{
\item{m}{nm object}
}
\description{
Gets attributes of the run like run time, queue time
}
