% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-code.R
\name{ls_scripts}
\alias{ls_scripts}
\title{List scripts}
\usage{
ls_scripts(folder = ".", extn = "r|R|Rmd|rmd", recursive = TRUE)
}
\arguments{
\item{folder}{String describing folder to search recursively in.}

\item{extn}{Character (can be regex) giving extension to limit search to.}

\item{recursive}{Logical (default = \code{TRUE}).  Should directories be searched
recursively.}
}
\value{
Character vector of matched file paths.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{

## find all scripts with the string "AUC("
ls_scripts("~/path/to/analysis/Scripts") \%>\% search_raw("AUC\\\\(")

}
