% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{make_boot_datasets}
\alias{make_boot_datasets}
\title{add/write bootstrap datasets}
\usage{
make_boot_datasets(
  m,
  samples = 10,
  data_folder = "DerivedData/bootstrap_datasets",
  overwrite = FALSE,
  id_var = "ID",
  ...
)
}
\arguments{
\item{m}{nm object}

\item{samples}{number of samples}

\item{data_folder}{folder to store datasets}

\item{overwrite}{overwrite or not}

\item{id_var}{character (default = "ID"). Name of ID column}

\item{...}{arguments passed to fill_input}
}
\description{
add/write bootstrap datasets
}
