% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{make_boot_datasets}
\alias{make_boot_datasets}
\title{add/write bootstrap datasets}
\usage{
make_boot_datasets(
  bootsplits,
  index = 1:10,
  m,
  data_folder = "DerivedData/bootstrap_datasets",
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{bootsplits}{rsplits dataset from \code{rsample::bootstrap()}}

\item{index}{index for subsetting}

\item{m}{nm object}

\item{data_folder}{folder to store datasets}

\item{overwrite}{overwrite or not}

\item{...}{arguments passed to fill_input}
}
\description{
preparation step before creating nm models
}
