% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_OCC_every_dose.R
\name{make_OCC_every_dose}
\alias{make_OCC_every_dose}
\title{Experiment - make an OCC column for NONMEM in dataset}
\usage{
make_OCC_every_dose(d, dose_trigger, new_OCC_trigger)
}
\arguments{
\item{d}{data.frame. NONMEM input dataset}

\item{dose_trigger}{logical expression for defining a dosing row}

\item{new_OCC_trigger}{logical expression for defining when OCC should increment}
}
\description{
To be used in a \code{mutate()} statement \code{group_by}'d by "ID"
}
