% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{manual_edit}
\alias{manual_edit}
\title{Perform manual edit of control file}
\usage{
manual_edit(m, description)
}
\arguments{
\item{m}{nm object}

\item{description}{character. Description of edit for documentation purposes}
}
\description{
Perform manual edit of control file
}
