% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctl_handling.R
\name{new_ctl}
\alias{new_ctl}
\title{make new control file based on previous}
\usage{
new_ctl(r, run_id)
}
\arguments{
\item{r}{object of class nmexecute, ctl_list, ctl_character or character}

\item{run_id}{character or numeric. new run_id}
}
\description{
make new control file based on previous
}
