% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctl_handling.R
\name{new_ctl}
\alias{new_ctl}
\title{make new control file based on previous}
\usage{
new_ctl(r, run_id, based_on, dir = getOption("models.dir"))
}
\arguments{
\item{r}{object coercible into ctl_list}

\item{run_id}{character or numeric. new run_id}

\item{based_on}{optional character new run_id}

\item{dir}{character. default "Models" dir}
}
\description{
make new control file based on previous
}
\examples{
\dontrun{
new_ctl(m1, "m2")  ## "models/runm2.mod"
new_ctl(m1, "models/runm2.mod") ## "models/runm2.mod"
new_ctl(m1, "runm2.mod")  ## "models/runm2.mod"
new_ctl(m1, "models/a/b/runm2.mod")  ## "models/a/b/runm2.mod"
}
}
