% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{nm}
\alias{nm}
\title{Make object of class nm}
\usage{
nm(cmd, psn_command, shell_script_name, shell_script_extn = "sh", ctl_name,
  run_id, run_dir, run_in = getOption("models.dir"),
  db_name = "runs.sqlite")
}
\arguments{
\item{cmd}{character. system command to launch NONMEM (PsN)}

\item{psn_command}{character. Name of PsN command (optional)}

\item{shell_script_name}{character. Name of shell script for grid submission (optional)}

\item{shell_script_extn}{character. File extension shell script for grid submission (optional)}

\item{ctl_name}{character. Name of control file (optional)}

\item{run_id}{character or numeric. Run identifier (optional)}

\item{run_dir}{character or numeric. Run directory (optional)}

\item{run_in}{character. Directory to execute. Default = getOption("models.dir")}

\item{db_name}{character. Name of db. NULL = don't put in database}
}
\value{
object of class nm
}
\description{
Make object of class nm
}
