% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{nm_create_analysis_project}
\alias{nm_create_analysis_project}
\title{Create analysis project}
\usage{
nm_create_analysis_project(
  path,
  dirs = nm_default_dirs(),
  style = c("simple"),
  use_renv = FALSE,
  ...
)
}
\arguments{
\item{path}{character path (relative or absolute) to project.  If just
specifying a name, this will create the analysis project in the current
working directory}

\item{dirs}{character list or vector.  Default = `nm_default_dirs()`}

\item{style}{character. Currently only "simple" See details}

\item{use_renv}{logical (default = FALSE). should renv be used or not in
project.}

\item{...}{deprecated}
}
\description{
Create analysis project
}
\details{
function works like `starters::create_analysis_project`.  With a few
  differences.  There is no restriction on directory name.  It is therefore
  possible to violate package naming conventions.
}
