% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMproject.R
\name{nm_default_dir}
\alias{nm_default_dir}
\title{Get a default directory}
\usage{
nm_default_dir(name = c("scripts", "models", "results"), ...)
}
\arguments{
\item{name}{Character. Directory type.  Should be either \code{"scripts"},
\code{"models"} or \code{"results"}.}

\item{...}{Not used.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
nm_default_dir("scripts")
nm_default_dir("models")
nm_default_dir("results")
}
