% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMproject_options.R
\name{nm_default_dirs}
\alias{nm_default_dirs}
\title{Setup analysis subdirectories}
\usage{
nm_default_dirs(dir_list)
}
\arguments{
\item{dir_list}{Optional named list or vector. Names \code{"scripts"} and
\code{"models"} must be present.  The rest can be unnamed.}
}
\value{
if \code{dir_list} is missing, will return value of
\code{getOption("nm_default_dirs")} otherwise will set option \code{nm_default_dirs}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This allows organisations/individuals with their own directory to customize
their directory structure
}
\examples{
\dontrun{

nm_default_dirs()
nm_default_dirs(list(models = "Models", 
                     scripts = "Scripts",
                     results = "Results",
                     "SourceData",
                     "Data"))
}
}
