% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{nm_diff}
\alias{nm_diff}
\title{Compute diff between two NONMEM runs}
\usage{
nm_diff(ref_m, m)
}
\arguments{
\item{ref_m}{nm object (base/reference object)}

\item{m}{nm object}
}
\value{
diff object
}
\description{
NMproject's control file manipulation functions (e.g. subroutine())
 may not work for all control files. It is the responsibilty of 
 the user to check automatic manipulations are done properly.
 Displaying diffs provides a means of manually checking.
}
\examples{
\dontrun{

m1 <- nm(run_id = "m1") \%>\%
  ctl("staging/Models/run1.mod")

m2 <- m1 \%>\% child(run_id = "m2") \%>\%
  subroutine(advan = 2, trans = 2)

nm_diff(m1, m2)

}
}
