% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_gettersetters.R
\name{nm_getsetters}
\alias{nm_getsetters}
\alias{run_dir}
\alias{cmd}
\alias{type}
\alias{parent_run_id}
\alias{parent_run_in}
\alias{parent_ctl_name}
\alias{parent_results_dir}
\alias{unique_id}
\alias{ctl_name}
\alias{results_dir}
\alias{cores}
\alias{parafile}
\alias{walltime}
\alias{executed}
\alias{lst_path}
\alias{run_in}
\alias{run_id}
\alias{result_files}
\title{functions to access and modify fields of nm objects}
\usage{
run_dir(m, text)

cmd(m, text)

type(m, text)

parent_run_id(m, text)

parent_run_in(m, text)

parent_ctl_name(m, text)

parent_results_dir(m, text)

unique_id(m, text)

ctl_name(m, text)

results_dir(m, text)

cores(m, text)

parafile(m, text)

walltime(m, text)

executed(m, text)

lst_path(m, text)

run_in(m, text)

run_id(m, text)

result_files(m, text)
}
\arguments{
\item{m}{nm object}

\item{text}{optional character for replacing field. If present function will
modify field (of same name as function) otherwise will return value of
field (of same name as function)}
}
\description{
functions to access and modify fields of nm objects
}
\details{
Easiest way to see all fields of an object is to type the object
  into the console and hit enter. This will display the value of each field.
  some fields like \code{cmd} are glue fields.  In these cases inserting
  expressions inside braces in \code{text} will evaluate the expression
}
\examples{
\dontrun{

## set cmd field of m1
m1 <- m1 \%>\% cmd("execute {ctl_name} -dir={run_dir}")

m1 \%>\% cmd() ## display command
## can also view field when viewing object
m1

}
}
