% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{nm_list_gather}
\alias{nm_list_gather}
\title{get all nm_list objects}
\usage{
nm_list_gather(envir = .GlobalEnv)
}
\arguments{
\item{envir}{environment (default = .GlobalEnv) to search}
}
\description{
get all nm_list objects
}
