% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{nm_list_render}
\alias{nm_list_render}
\title{render function for child nm_lists}
\usage{
nm_list_render(
  m,
  input,
  output_file = NA,
  args = list(),
  force = getOption("nm.force_render"),
  async = FALSE,
  ...
)
}
\arguments{
\item{m}{nm object}

\item{input}{character. Same as rmarkdown::render() arg}

\item{output_file}{character. Same as rmarkdown::render() arg}

\item{args}{list. Same as "params" arg in rmarkdown::render()}

\item{force}{logical (default = FALSE). will force execution}
}
\description{
Mostly used for bootstraps, and other routines where
  a parent run spawns multiple children in the form of an nm_list
}
