% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_table.R
\name{nm_output}
\alias{nm_output}
\title{Get NONMEM output tables}
\usage{
nm_output(r, dorig, ...)
}
\arguments{
\item{r}{An nm object.}

\item{dorig}{Optional \code{data.frame}. NONMEM input dataset.}

\item{...}{Additional arguments to pass on to \code{\link[=read.csv]{read.csv()}}.}
}
\value{
A list of \code{tibble}s with merged version of all output $TABLEs and the
input data.  Additional columns will be \code{INNONMEM} which will be TRUE for
rows that were not ignored by NONMEM.  For simulation control files there
is also \code{DV_OUT} which will contain simulated \code{DV} values. \code{DV} will always
be unmodified from the input dataset.
}
\description{
This function may become deprecated soon. The function \code{\link[=output_table]{output_table()}} is a
more user friendly wrapper for this function.
}
\keyword{internal}
