% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{nm_output}
\alias{nm_output}
\title{Get NONMEM output tables}
\usage{
nm_output(r, dorig, ...)
}
\arguments{
\item{r}{data.frame.  object of class nmexecute}

\item{dorig}{data.frame. optional NONMEM input dataset.}

\item{...}{additional arguments to pass on to read.csv}
}
\description{
This combines $TABLE output with the input data, allowing text columns to be retained for plotting/summaries.
}
