% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{nm_output_path}
\alias{nm_output_path}
\title{Find an output file}
\usage{
nm_output_path(m, extn, file_name)
}
\arguments{
\item{m}{nm object}

\item{extn}{character. name of extension}

\item{file_name}{optional character. name of file name}
}
\description{
Find an output file
}
\examples{
\dontrun{
m \%>\% nm_output_path("ext")  ## path to ext file
}
}
