% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-run-summaries.R
\name{nm_output_path}
\alias{nm_output_path}
\title{Find an output file associated with a run}
\usage{
nm_output_path(m, extn, file_name)
}
\arguments{
\item{m}{An nm object.}

\item{extn}{Character. Name of extension.}

\item{file_name}{Optional character. Name of file name.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This is primarily a backend function used to identify output file paths
associated with nm objects.
}
\examples{
\dontrun{
m \%>\% nm_output_path("ext")  ## path to ext file
}
}
