% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_read_table.R
\name{nm_read_table}
\alias{nm_read_table}
\title{Fast read of NONMEM output table}
\usage{
nm_read_table(file, ...)
}
\arguments{
\item{file}{File argument from \code{utils::read.table()}.}

\item{...}{Other arguments to be passed to \code{utils::read.table()}.}
}
\value{
A \code{data.frame} from a relevant $TABLE output file.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Reads in $TABLE outputs rapidly.  \code{\link[=output_table]{output_table()}} is a higher level function
for reading output files and combining with input datasets.
}
\seealso{
\code{\link[=output_table]{output_table()}}
}
