% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{nm_row}
\alias{nm_row}
\title{convert nm objects to a rowwise tibble/data.frame}
\usage{
nm_row(m)
}
\arguments{
\item{m}{nm object}
}
\description{
used mainly internally, but this could find use with advanced users
}
