% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_object.R
\name{nm_row}
\alias{nm_row}
\title{Convert nm objects to a rowwise tibble/data.frame}
\usage{
nm_row(m)
}
\arguments{
\item{m}{An nm object.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Primarily an internal function. Converts an \code{nm_list} object to a \code{tibble}.
}
\keyword{internal}
