% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_tran.R
\name{nm_tran}
\alias{nm_tran}
\title{Run NMTRAN step only}
\usage{
nm_tran(x)
}
\arguments{
\item{x}{nm object}
}
\description{
Runs initial part of NONMEM where control file and dataset checks are performed.
Stops before running NONMEM.  Useful especially on grid infrastructures where 
it may take a while for NONMEM to start return ctl/dataset errors
}
\details{
Requires \code{options("nmtran_exe_path")} to be set.
}
