% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMproject.R
\name{nm_tran_command}
\alias{nm_tran_command}
\title{Get/set nm_tran_command}
\usage{
nm_tran_command(text)
}
\arguments{
\item{text}{Optional character. If specified will set \code{nm_tran_command}.}
}
\value{
if \code{text} is missing will get and return the current NMTRAN command
}
\description{
\code{nm_tran} needs the location of NMTRAN.exe on your system.  This
is guessed on package load, assuming PsN is on the $PATH environmental
variable, if this is not the case, then you can manually set the path and
command used.
}
\details{
\code{text} can just be the path to NMTRAN.exe.  In this case
\code{nm_tran_command} will use the format \verb{/path/to/NMTRAN.exe < \{ctl_name\}} to
launch NMTRAN.exe where \code{{ctl_name}} is the name of the control file \code{text}
can use the \code{{ctl_name}} glue field however for more complicated commands.
See examples

Set this up either at the beginning of your script, in your \code{.Rprofile} or
for all users in \code{Rprofile.site}
}
\examples{
\dontrun{

  # the following two are equivalent
  nm_tran_command("/opt/NONMEM/nm75/tr/NMTRAN.exe")
  nm_tran_command("/opt/NONMEM/nm75/tr/NMTRAN.exe < {ctl_name}")

}
}
\seealso{
\code{\link[=find_nm_tran_path]{find_nm_tran_path()}}
}
