% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctl_handling.R
\name{nonmem_code_to_r}
\alias{nonmem_code_to_r}
\title{convert nonmem code to R ready}
\usage{
nonmem_code_to_r(code, eta_to_0 = TRUE)
}
\arguments{
\item{code}{character vector of NONMEM code block}

\item{eta_to_0}{logical (default = TRUE) set all etas to 0}
}
\description{
This is a developer function
}
