% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctl_handling.R
\name{nonmem_code_to_r}
\alias{nonmem_code_to_r}
\title{Convert nonmem code to R ready}
\usage{
nonmem_code_to_r(code, eta_to_0 = TRUE)
}
\arguments{
\item{code}{Character vector of NONMEM code block.}

\item{eta_to_0}{Logical (default = \code{TRUE}). Should all etas be set to 0.}
}
\value{
Character vector of R code.
}
\description{
Parses NONMEM code and attempts to make it evaluable as R code.
}
\keyword{internal}
