% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nonmem.R
\name{NONMEM_version}
\alias{NONMEM_version}
\title{NONMEM version info}
\usage{
NONMEM_version()
}
\value{
Returns list with version info for NONMEM, PsN, perl and
fortran compiler (only gfortran currently)
}
\description{
Gets version information about the NONMEM installation, PsN installation and compilers.
}
