% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofv.R
\name{ofv}
\alias{ofv}
\title{Get Objective Function Value (OFV)}
\usage{
ofv(r)
}
\arguments{
\item{r}{An nm object.}
}
\value{
The \code{numeric} value of the OFV.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Extracts OFV from .ext file.
}
\examples{

# create example object m1 from package demo files
exdir <- system.file("extdata", "examples", "theopp", package = "NMproject")
m1 <- new_nm(run_id = "m1", 
             based_on = file.path(exdir, "Models", "ADVAN2.mod"),
             data_path = file.path(exdir, "SourceData", "THEOPP.csv"))

is_finished(m1) ## FALSE until run is completed
ofv(m1) ## NA until m1 is finished

}
