% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofv.R
\name{ofv}
\alias{ofv}
\title{Get objective function value (OFV)}
\usage{
ofv(r)
}
\arguments{
\item{r}{An nm object.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
\dontrun{

ofv(m1)

}
}
