% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_record.R
\name{output_table}
\alias{output_table}
\alias{output_table_first}
\title{Get processed output table}
\usage{
output_table(r, only_append = c(), ...)

output_table_first(r, ...)
}
\arguments{
\item{r}{object of class nm}

\item{only_append}{optional character vector. If missing will append all,
otherwise will append only those variables requested}

\item{...}{optional additional arguments to pass on to read.csv of orig data}
}
\value{
\code{output_table} will return a list of tibbles with merged version
  of all output $TABLEs and the input data.  Additional columns will be
  \code{INNONMEM} which will be TRUE for rows that were not ignored by
  NONMEM.  For simulation control files there is also \code{DV_OUT} which
  will contain simulated \code{DV} values. \code{DV} will always be
  unmodified from the input dataset

\code{output_table_first} will return a tibble with a single run.
}
\description{
Get processed output table
}
