% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMproject.R
\name{overwrite_behaviour}
\alias{overwrite_behaviour}
\title{Overwrite behaviour of NMproject}
\usage{
overwrite_behaviour(txt = c("ask", "overwrite", "stop", "skip"))
}
\arguments{
\item{txt}{character either "run", "stop", "skip"}
}
\description{
Requires setting "nm.overwrite_behaviour" \code{option()}
}
