% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{param_cov_diag}
\alias{param_cov_diag}
\title{plot relationship between a parameter and covariate}
\usage{
param_cov_diag(r, param, cov, ..., categorical = FALSE, plot_tv = TRUE)
}
\arguments{
\item{r}{nm object}

\item{param}{character. Name of parameter}

\item{cov}{character. Name of covariate}

\item{...}{additional arguments passed to dplyr::mutate()}

\item{categorical}{logical (default = FALSE)}

\item{plot_tv}{logical}
}
\description{
plot relationship between a parameter and covariate
}
\details{
the mutate statement is to add variables not included in original $TABLE
}
