% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low-level-ctl-handling-funs.R
\name{param_info}
\alias{param_info}
\title{Get parameter information}
\usage{
param_info(ctl)
}
\arguments{
\item{ctl}{Character. Path to control file.}
}
\description{
Get parameter information
}
\keyword{internal}
