% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_object.R
\name{parent_run}
\alias{parent_run}
\title{Get parent object of nm object}
\usage{
parent_run(m, n = 1L)
}
\arguments{
\item{m}{nm object}

\item{n}{numeric. generation of parent (default = 1)}
}
\description{
Get parent object of nm object
}
