% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_iter.R
\name{plot_iter}
\alias{plot_iter}
\title{Plot iterations vs parameters/OBJ}
\usage{
plot_iter(r, trans = TRUE, skip = 0, yvar = "OBJ")
}
\arguments{
\item{r}{An nm object.}

\item{trans}{Logical (default = \code{TRUE}). Should parameter transformations be
performed in accordance with $THETA/$OMEGA/$SIGMA comments}

\item{skip}{Numeric (default = 0). The number of iterations to skip when
plotting.  For observing stationarity it is often useful to remove the
beginning iterations where the OFV and parameters may move a lot.}

\item{yvar}{Character (default = \code{"OBJ"}). Name of variable/parameter to
display.}
}
\value{
A \code{ggplot2} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Non interactive \code{ggplot2} based version of the OFV/parameter vs iteration
plot in shiny run monitor \code{\link[=shiny_nm]{shiny_nm()}}.  Used mainly for inclusion in
diagnostic reports.
}
\seealso{
\code{\link[=shiny_nm]{shiny_nm()}}, \code{\link[=nm_render]{nm_render()}}
}
