% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_iter.R
\name{plot_iter_data}
\alias{plot_iter_data}
\title{Get data for iterations vs parameters/OBJ plots}
\usage{
plot_iter_data(r, trans = TRUE, skip = 0, yvar = "OBJ")
}
\arguments{
\item{r}{object of class nmexecute}

\item{trans}{logical. should parameter transformations be performed}

\item{skip}{numeric (default 0). how many iterations to skip when plotting}

\item{yvar}{character (default = "OBJ"). Name of variable/parameter to display}
}
\description{
Get data for iterations vs parameters/OBJ plots
}
