% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-apps.R
\name{plot_iter_dygraph}
\alias{plot_iter_dygraph}
\title{Plot iterations vs parameters/OBJ with dygraphs}
\usage{
plot_iter_dygraph(m, trans = TRUE)
}
\arguments{
\item{m}{An nm object}

\item{trans}{Logical (default = \code{TRUE}). Should parameter transformations be
performed in accordance with $THETA/$OMEGA comments.}
}
\value{
A \code{dygraph} plot.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\keyword{internal}
