% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_iter.R
\name{plot_iter_ggplot}
\alias{plot_iter_ggplot}
\title{Plot iterations vs parameters/OBJ (ggplot2)}
\usage{
plot_iter_ggplot(d)
}
\arguments{
\item{d}{data.frame. output from plot_iter_data}
}
\description{
Plot iterations vs parameters/OBJ (ggplot2)
}
