% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_reestExtra.R
\name{ppc_data}
\alias{ppc_data}
\title{process data from PPC for plotting}
\usage{
ppc_data(r, FUN, statistic = "statistic", ...)
}
\arguments{
\item{r}{nm object}

\item{FUN}{statistic function with NONMEM dataset as arg and
returns data.frame with a column "statistic"}

\item{statistic}{character (default = "statistic") name of statistic column
returned by FUN}

\item{additional}{arguments passed to dplyr::mutate for data manipulaion prior
to statistic (FUN) calculation}
}
\description{
process data from PPC for plotting
}
