% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_reestExtra.R
\name{ppc_data}
\alias{ppc_data}
\title{process data from PPC for plotting}
\usage{
ppc_data(r, FUN, ..., pre_proc = identity, max_mod_no = NA, DV = "DV",
  statistic = "statistic")
}
\arguments{
\item{r}{nm object}

\item{FUN}{statistic function with NONMEM dataset as arg and
returns data.frame with a column "statistic"}

\item{...}{additional arguments for FUN}

\item{pre_proc}{function to mutate dataset prior to compute statistics}

\item{DV}{character (default = "DV")}

\item{statistic}{character (default = "statistic") name of statistic column
returned by FUN}

\item{max_mod_mo}{integer. Maximum model number to read (set low for debugging)}
}
\description{
process data from PPC for plotting
}
\examples{
\dontrun{

idEXPstat <- function(d, ...){ ## example statistic function

d \%>\% group_by(ID, ...) \%>\% filter(is.na(AMT)) \%>\%
 summarise(
   AUC = AUC(time = TIME, conc = DV),
   CMAX = max(DV, na.rm = TRUE),
   TMAX = TIME[which.max(DV)]
 ) \%>\%
 tidyr::gather(key = "exposure", value = "statistic", AUC:TMAX) \%>\%
 ungroup()
}

EXPstat <- function(d, ...){ ## example statistic function
 ## arg = nonmem dataset data.frame
 ## return data.frame with statistic column
 d \%>\% idEXPstat(...) \%>\%  ## reuse idEXPstat for individual stats
   ## summarise over study and any other variables (...)
   group_by(exposure, ...) \%>\%
   summarise(
   median = median(statistic, na.rm = TRUE),
   cv = 100*sd(statistic, na.rm = TRUE)/mean(statistic, na.rm = TRUE)
 ) \%>\%
 tidyr::gather(key = "type", value = "statistic", median:cv)
}

m1s \%>\% ppc_data(EXPstat)

}
}
\seealso{
\code{\link(ppc_whisker_plot)}, \code{\link(ppc_histogram_plot)}
}
