% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{preview}
\alias{preview}
\title{Preview code_library file}
\usage{
preview(name)
}
\arguments{
\item{name}{character indicating script in code_library to preview}
}
\description{
Preview code_library file
}
