% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psn_style_scm.R
\name{psn_style_scm}
\alias{psn_style_scm}
\title{PsN style stepwise covariate method}
\usage{
psn_style_scm(base, run_in, dtest, alpha_forward = 0.05, alpha_backward = 0.01)
}
\arguments{
\item{base}{An nm object (base model).}

\item{run_in}{A directory to run in.}

\item{dtest}{Output of \code{\link[=test_relations]{test_relations()}}.}

\item{alpha_forward}{Numeric (default = \code{0.05}). Alpha level for forward
inclusion.}

\item{alpha_backward}{Numeric (default = \code{0.01}). Alpha level for backward
deletion.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Intent is not to replicate PsN SCM.  This is mainly here for illustrative and
comparison purposes. Should replicate the model selection in PsN's SCM
functionality with greedy setting.
}
\seealso{
\code{\link[=test_relations]{test_relations()}}, \code{\link[=covariate_step_tibble]{covariate_step_tibble()}},
\code{\link[=bind_covariate_results]{bind_covariate_results()}}.
}
