% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{read_derived_data}
\alias{read_derived_data}
\title{Read derived data}
\usage{
read_derived_data(name, na = ".", ...)
}
\arguments{
\item{name}{name of file (without extension)}

\item{na}{character to be passed to read.csv}

\item{...}{additional arguments to be passed to read.csv}
}
\description{
Read derived data
}
