% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{relative_path}
\alias{relative_path}
\title{Compute path relative to reference}
\usage{
relative_path(path, relative_path)
}
\arguments{
\item{path}{Path of desired directory or file.}

\item{relative_path}{Path of directory to travel from.}
}
\value{
A relative path \code{character} object.
}
\description{
Compute path relative to reference
}
\keyword{internal}
