% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctl_handling.R
\name{remove_cov}
\alias{remove_cov}
\title{Remove a covariate to a NONMEM model}
\usage{
remove_cov(
  ctl,
  param,
  cov,
  state = 2,
  continuous = TRUE,
  time_varying,
  additional_state_text,
  id_var = "ID"
)
}
\arguments{
\item{ctl}{object coercible to ctl_list}

\item{param}{character. Name of parameter}

\item{cov}{character. Name of covariate}

\item{state}{numeric or character. Number/name of state (see details)}

\item{continuous}{logical (default = TRUE). is covariate continuous?}

\item{time_varying}{optional logical. is the covariate time varying?}

\item{additional_state_text}{optional character. custom state variable to be passed to param_cov_text}

\item{id_var}{character (default = "ID"). Needed if time_varying is missing.}
}
\description{
Only works with covariates added with add_cov
}
\details{
available states:
"2" or "linear":
  PARCOV= ( 1 + THETA(1)*(COV - median))

"3" or "hockey-stick":
  IF(COV.LE.median) PARCOV = ( 1 + THETA(1)*(COV - median))
  IF(COV.GT.median) PARCOV = ( 1 + THETA(2)*(COV - median))
                    
"4" or "exponential":
   PARCOV= EXP(THETA(1)*(COV - median))

"5" or "power":
   PARCOV= ((COV/median)**THETA(1))
   
"power1":
   PARCOV= ((COV/median))
   
"power0.75":
   PARCOV= ((COV/median)**0.75)

"6" or "log-linear":
   PARCOV= ( 1 + THETA(1)*(LOG(COV) - log(median)))
}
\examples{
\dontrun{

m1noWT <- m1 \%>\% child("m1noWT") \%>\%
  remove_cov(param = "CL", cov = "WT") \%>\% 
  run_nm()
  
## compare results

rr(c(m1, m1noWT))
summary_wide(c(m1, m1noWT))
  
}

}
\seealso{
\code{\link{covariate_step_tibble}}
}
