% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_object.R
\name{remove_parameter}
\alias{remove_parameter}
\title{remove parameter from NONMEM control file}
\usage{
remove_parameter(m, name)
}
\arguments{
\item{m}{nm object}

\item{name}{character. Parameter to remove}
}
\description{
remove parameter from NONMEM control file
}
