% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_object.R
\name{rename_parameter}
\alias{rename_parameter}
\title{rename a parameter in NONMEM control stream}
\usage{
rename_parameter(m, ...)
}
\arguments{
\item{m}{nm object}

\item{...}{renaming arguments}
}
\description{
rename a parameter in NONMEM control stream
}
\examples{
\dontrun{

m1 \%>\% rename_parameter("CL" = "KE")
## renames KE to CL

}
}
