% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{run}
\alias{run}
\title{Run NONMEM}
\usage{
run(..., overwrite = .sso_env$run_overwrite, delete_dir = c(NA, TRUE,
  FALSE), wait = .sso_env$wait)
}
\arguments{
\item{...}{objects of class nm}

\item{overwrite}{logical. Should run directory be overwritten (default=FALSE)}

\item{delete_dir}{logical. NA (default - directory will be deleted if no dependencies exists)
TRUE or FALSE. Should run_dir be deleted.}

\item{wait}{logical (default=FALSE). Should R wait for run to finish.
Default can be changed with  wait_by_default() function}
}
\description{
Run NONMEM
}

