% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{run}
\alias{run}
\title{Run NONMEM}
\usage{
run(..., overwrite = .sso_env$run_overwrite, delete_dir = c(NA, TRUE,
  FALSE), wait = .sso_env$wait, update_db = TRUE, ignore.stdout = TRUE,
  ignore.stderr = TRUE, initial_timeout = NA, quiet = .sso_env$quiet)
}
\arguments{
\item{...}{objects of class nm}

\item{overwrite}{logical. Should run directory be overwritten (default=FALSE)}

\item{delete_dir}{logical. NA (default - directory will be deleted if no dependencies exists)
TRUE or FALSE. Should run_dir be deleted.}

\item{wait}{logical (default=FALSE). Should R wait for run to finish.
Default can be changed with  wait_by_default() function}

\item{update_db}{logical (default=TRUE). Should run_status be updated}

\item{ignore.stdout}{logical (default=TRUE). Parameter passed to system()}

\item{ignore.stderr}{logical (default=TRUE). Parameter passed to system()}

\item{initial_timeout}{numeric. time in seconds.
time period to give up on a run if directory hasn't been created.}

\item{quiet}{logical (default=FALSE). should system_nm output be piped to screen}
}
\description{
Run NONMEM
}
