% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{run}
\alias{run}
\title{Load nm object into memory from script}
\usage{
run(..., overwrite = FALSE)
}
\arguments{
\item{...}{objects of class nm}

\item{overwrite}{logical. Should run directory be overwritten (default=FALSE)}
}
\description{
Run NONMEM
}

