% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-management.R
\name{run_all_scripts}
\alias{run_all_scripts}
\title{Run all project scripts sequentially}
\usage{
run_all_scripts(index, quiet = FALSE)
}
\arguments{
\item{index}{Numeric index for subsetting list of scripts before running.}

\item{quiet}{Argument passed to \code{\link[rmarkdown:render]{rmarkdown::render()}}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Runs/renders all scripts \code{s01_XXX}, \code{s02_XXX} in the designated "scripts"
directory.
}
\details{
Works with .R and .Rmd extensions.  Behaviour is to \code{\link[=source]{source()}} .R
files and use \code{\link[rmarkdown:render]{rmarkdown::render()}} on .Rmd files
}
