% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{run_all_scripts}
\alias{run_all_scripts}
\title{Run all project scripts sequentially}
\usage{
run_all_scripts(quiet = FALSE)
}
\arguments{
\item{quiet}{Argument passed to \code{rmarkdown::render}}
}
\description{
Runs all scripts s01_XXX, s02_XXX in the designated "scripts" directory
}
\details{
works with .R and .Rmd extensions.  Behaviour is to \code{source} .R
files and use \code{rmarkdown::render} on .Rmd files
}
