% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{run_dir_path}
\alias{run_dir_path}
\title{get path to run_dir}
\usage{
run_dir_path(m)
}
\arguments{
\item{m}{nm object}
}
\description{
get path to run_dir
}
\seealso{
\code{\link{nm_getsetters}}
}
