% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMproject.R
\name{run_id}
\alias{run_id}
\title{Get run id}
\usage{
run_id(m, text)
}
\arguments{
\item{m}{character or nm or ctl_list/ctl_character}

\item{text}{optional character to set run_id}

\item{...}{additional arguments}
}
\description{
Get run id
}
