% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_nm.R
\name{run_nm}
\alias{run_nm}
\title{Run NONMEM jobs}
\usage{
run_nm(
  m,
  wait = getOption("wait"),
  ignore.stdout = TRUE,
  ignore.stderr = TRUE,
  quiet = getOption("quiet_run"),
  intern = getOption("intern"),
  force = FALSE,
  cache_ignore_cmd = FALSE,
  cache_ignore_ctl = FALSE,
  cache_ignore_data = FALSE
)
}
\arguments{
\item{m}{nm objects}

\item{wait}{logical (default=FALSE). Should R wait for run to finish.
Default can be changed with  wait_by_default() function}

\item{ignore.stdout}{logical (default=TRUE). Parameter passed to system()}

\item{ignore.stderr}{logical (default=TRUE). Parameter passed to system()}

\item{quiet}{logical (default=FALSE). should system_nm output be piped to screen}

\item{intern}{logical. intern arg to be passed to system}

\item{force}{logical (default = FALSE).  Force run even results unchanged}

\item{cache_ignore_cmd}{logical (default = FALSE). Should check \code{cmd} field with cache?}

\item{cache_ignore_ctl}{logical (default = FALSE). Should check control file contents with cache?}

\item{cache_ignore_data}{logical (default = FALSE). Should check dataset with cache?}
}
\value{
m with \code{job_info} fields populated.
}
\description{
Run nm objects.  Uses \code{system_nm()} to submit the \code{cmd()} value of object
}
