% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_record.R
\name{run_record}
\alias{run_record}
\title{Run record}
\usage{
run_record(..., trans = TRUE)
}
\arguments{
\item{...}{objects of class nmexecute}

\item{trans}{logical (default=TRUE).
If possible TRUE means transformation will be attempted}
}
\description{
Run record
}
