% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-management.R
\name{setup_nm_demo}
\alias{setup_nm_demo}
\title{Setup demo in current directory}
\usage{
setup_nm_demo(
  demo_name = "theopp",
  overwrite = FALSE,
  additional_demo_locations = NULL
)
}
\arguments{
\item{demo_name}{Character. Name of demo. Default = "theopp".  See details to find other demos}

\item{overwrite}{Logical. Default changed to FALSE.}

\item{additional_demo_locations}{Character vector. default = NULL. Locations
for demo directories.}
}
\value{
Invisibly returns a \code{tibble} with imported file information.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Following through the demo is the fastest way to learn the syntax of
NMproject. The default demo is a Theophylline ("theopp") pharmacometric analysis.  Scripts will
be copied numbered \code{s01_XXX.Rmd}, \code{s02_XXX.Rmd} in the \code{"Scripts"} directory and a
dataset into \code{"SourceData"}. The \code{"staging"} area will also be pre-filled with
the code library model, \code{"ADVAN2.mod"}.  To practice copying this yourself, see
\code{\link[=code_library]{code_library()}} for how the app works.
}
\details{
Available \code{demo_name} correspond to directory locations in
\code{system.file("extdata","examples",package = "NMproject")}
}
\seealso{
\code{\link[=code_library]{code_library()}}
}
