% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{setup_nm_demo}
\alias{setup_nm_demo}
\title{Setup demo files}
\usage{
setup_nm_demo(demo_name = "theopp", new_project = NA,
  file_stub = paste0(getOption("model_file_stub"), 1), overwrite = FALSE,
  exclude = NULL, additional_demo_locations = NULL)
}
\arguments{
\item{demo_name}{character. Name of demo. Default = "theopp"}

\item{new_project}{character. To set up demo in a separate (new) project}

\item{file_stub}{character. Default = "run1". Stub to some file names}

\item{overwrite}{logical. Default changed to FALSE.}

\item{exclude}{character. Name of extension to exclude from copying}

\item{additional_demo_locations}{character vector. default = NULL.
locations for demo directories}
}
\description{
Setup demo files
}
