% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_apps.R
\name{shiny_nm}
\alias{shiny_nm}
\title{Run monitor & summary app}
\usage{
shiny_nm(m, envir = .GlobalEnv)
}
\arguments{
\item{m}{Either an nm object, or data.frame or list or environment contain
nm_lists.}

\item{envir}{If missing, the environment to search.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Interactively monitor NONMEM runs.  This interface is intentionally limited
to monitoring runs, and does not include the ability to create, modify,
launch or post-process runs since actions performed in the shiny app are not
traceable/reproducible and not part of the workflow you create when
scripting.
}
\examples{
\dontrun{

shiny_nm()        ## use all objects in global workspace
shiny_nm(m1)      ## only m1
shiny_nm(d$m)     ## only d$m
shiny_nm(d)       ## all nm_lists in d (data.frame/list/environment)

}
}
