% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-file.R
\name{show_ctl}
\alias{show_ctl}
\title{Show an uneditable version of the control file}
\usage{
show_ctl(r)
}
\arguments{
\item{r}{An nm object.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Opens a read-only version of the NONMEM control file for browsing.
}
\seealso{
\code{\link[=show_out]{show_out()}}.
}
