% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{stage}
\alias{stage}
\title{Stage files in project staging area ready for import}
\usage{
stage(files, root_dir, overwrite = FALSE, silent = FALSE)
}
\arguments{
\item{files}{character vector. path of files to stage}

\item{root_dir}{character path to root directory of `files`. Staged files
relative to `staging` directory will be same as `files` to `root_dir`. If
this is not specified, will guess based on presense of `nm_default_dirs`}

\item{overwrite}{logical (default = FALSE).}

\item{silent}{logical (default = FALSE)}
}
\description{
Stage files in project staging area ready for import
}
