% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_edit.R
\name{start_manual_edit}
\alias{start_manual_edit}
\title{internal function: start manual edit}
\usage{
start_manual_edit(m, combine_patch = NA_character_)
}
\arguments{
\item{m}{nm object}

\item{combine_patch}{logical. should patch be added first before edit}
}
\description{
Needed for addin functionality
}
