% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring.R
\name{status_table}
\alias{status_table}
\title{Get status of multiple runs in form of table}
\usage{
status_table(m)
}
\arguments{
\item{m}{An nm object.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

A more friendly version of \code{\link[=status]{status()}} for vector valued nm objects.  Useful
after bootstraps, or stepwise covariate method steps, or any situation
dealing with groups of NONMEM runs.
}
