% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{subroutine}
\alias{subroutine}
\title{Subroutine}
\usage{
subroutine(m, advan = NA, trans = 1, recursive = TRUE)
}
\arguments{
\item{m}{nm object}

\item{advan}{character. desired ADVAN}

\item{trans}{character. desired TRANS}

\item{recursive}{logical (default = TRUE). Internal argument, do not modify}
}
\description{
This is currently experimental. Set/modify subroutine of an nm object
}
\details{
Can only switch between subroutines listed in \code{available_advans}
 TODO: modify parameter initial estimates based on conversion
}
\examples{
\dontrun{

advan(m)  ## 2
trans(m)  ## 1

m <- m \%>\% subroutine(ADVAN = 2, TRANS = 2)

}
}
