% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subroutine.R
\name{subroutine}
\alias{subroutine}
\title{Subroutine}
\usage{
subroutine(m, advan = NA, trans = 1, recursive = TRUE)
}
\arguments{
\item{m}{nm object}

\item{advan}{character. desired ADVAN}

\item{trans}{character. desired TRANS}

\item{recursive}{logical (default = TRUE). Internal argument, do not modify}
}
\description{
This is currently experimental. Modify subroutine of an nm object
}
\details{
Can only switch between subroutines listed in \code{available_advans}
}
\examples{
\dontrun{

advan(m)  ## 2
trans(m)  ## 1

m <- m \%>\% subroutine(ADVAN = 2, TRANS = 2)

ds <- .available_advans \%>\%
  filter(oral) \%>\%
  mutate(
      m = m1 \%>\% child(run_id = label) \%>\%
        subroutine(advan = advan, trans = trans)
  )

}
}
\seealso{
\code{\link[=advan]{advan()}}
}
