% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{summary_wide}
\alias{summary_wide}
\title{#' @export
summary_wide <- function(..., parameters = c("none", "new", "all")){
  UseMethod("summary_wide")
}}
\usage{
summary_wide(..., parameters = c("none", "new", "all"), m = TRUE,
  trans = TRUE)
}
\description{
#' @export
summary_wide <- function(..., parameters = c("none", "new", "all")){
  UseMethod("summary_wide")
}
}
