% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{system_cmd}
\alias{system_cmd}
\title{System/shell command wrapper}
\usage{
system_cmd(cmd, dir = ".", ...)
}
\arguments{
\item{cmd}{character. command to send to shell}

\item{dir}{character. directory to run command in}

\item{...}{other arguments passed to system command}
}
\description{
System/shell command wrapper
}
