% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMproject.R
\name{system_nm}
\alias{system_nm}
\title{System command for NONMEM execution}
\usage{
system_nm(cmd, dir = nm_default_dir("models"), ...)
}
\arguments{
\item{cmd}{character. command to send to shell}

\item{dir}{character. directory to run command in}

\item{...}{other arguments passed to system command}
}
\description{
System command for NONMEM execution
}
