% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{target}
\alias{target}
\alias{untarget}
\title{Target part of control object for further modification}
\usage{
target(m, dollar, lines)

untarget(m, dollar)
}
\arguments{
\item{m}{An nm object.}

\item{dollar}{Character. Name of subroutine to target.}

\item{lines}{Optional character.  Assignment of lines.}
}
\value{
An nm object with modified \code{target} field.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Intended mostly for internal use.  Used with \code{\link[=text]{text()}} to target control file
modifications to specific NONMEM subroutines
}
\keyword{internal}
