% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{uncomment}
\alias{uncomment}
\title{uncomment lines of control file}
\usage{
uncomment(m, pattern = ".*")
}
\arguments{
\item{m}{nm object}

\item{pattern}{optional character regex.  Passed to gsub}
}
\description{
uncomment lines of control file
}
