% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctl_handling.R
\name{update_dollar}
\alias{update_dollar}
\title{Update subroutine}
\usage{
update_dollar(ctl, ..., which_dollar = 1, append = FALSE)
}
\arguments{
\item{ctl}{object coercible to ctl_list}

\item{...}{named replacement argument of class character}

\item{which_dollar}{(default = 1) which subroutine if multiple matches}

\item{append}{logical (default = FALSE).  Should results be appended to subroutine}
}
\description{
Update subroutine
}
\examples{
\dontrun{
m1 \%>\% update_dollar(THETA=c("1             	; KA ; h-1 ; LOG
                             -2.5            ; K  ; h-1 ; LOG
                             -0.5          	; V  ; L ; LOG")) \%>\%
       write_ctl()
}
}
