% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_object.R
\name{update_parameters}
\alias{update_parameters}
\title{update parameters from a control stream}
\usage{
update_parameters(ctl, from)
}
\arguments{
\item{ctl}{object coercible into ctl_list}

\item{from}{class nm. object from which to extract results}
}
\description{
update parameters from a control stream
}
