% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{valid_package_name}
\alias{valid_package_name}
\title{Package name validator from usethis}
\usage{
valid_package_name(x)
}
\arguments{
\item{x}{name of package}
}
\description{
Package name validator from usethis
}
\keyword{internal}
