% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_edit.R
\name{view_patch}
\alias{view_patch}
\title{View a patch}
\usage{
view_patch(patch_name)
}
\arguments{
\item{patch_name}{Character. Name of patch}
}
\description{
Use the "view patch" RStudio 'Addin' for viewing patches instead.
}
\keyword{internal}
