% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_edit.R
\name{view_patch}
\alias{view_patch}
\title{View a patch}
\usage{
view_patch(patch_name)
}
\arguments{
\item{patch_name}{character. name of patch}
}
\description{
Internal function. Use the addin for viewing patches instead
}
