% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring.R
\name{wait_finish}
\alias{wait_finish}
\title{Wait for runs to finish}
\usage{
wait_finish(r, timeout = NA)
}
\arguments{
\item{r}{nm object}

\item{timeout}{numeric seconds to wait before timeout}
}
\description{
blocks subsequent r execution until run(s) are finished
}
