% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{wait_finish}
\alias{wait_finish}
\title{wait for runs to finish}
\usage{
wait_finish(r, timeout = NA, force = FALSE)
}
\arguments{
\item{r}{nm object}

\item{timeout}{numeric seconds to wait before timeout}
}
\description{
blocks subsequent r execution until run(s) are finished
}
