% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{wait_for}
\alias{wait_for}
\title{Wait for statement to be true}
\usage{
wait_for(x, timeout = NULL)
}
\arguments{
\item{x}{expression to evaluate}

\item{timeout}{numeric. Maximum time (seconds) to wait}
}
\description{
Wait for statement to be true
}

