% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{wait_for_finished}
\alias{wait_for_finished}
\title{wait for a run to finish}
\usage{
wait_for_finished(..., initial_timeout = NA)
}
\arguments{
\item{...}{objects of class nm}

\item{initial_timeout}{numeric. time in seconds.
time period to give up on a run if directory hasn't been created.}
}
\description{
wait for a run to finish
}
