% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{write.csv.nm}
\alias{write.csv.nm}
\title{write csv for NONMEM control files}
\usage{
write.csv.nm(..., na = ".", row.names = FALSE, quote = FALSE)
}
\arguments{
\item{...}{arguments for write.csv}

\item{na}{character. Default changed to ".".}

\item{row.names}{logical. Default changed to FALSE.}

\item{quote}{logical. Fefault changed to FALSE}
}
\description{
write csv for NONMEM control files
}
