% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_nm.R
\name{write_ctl}
\alias{write_ctl}
\title{Write control file to disk}
\usage{
write_ctl(m, path = NA_character_, force = FALSE)
}
\arguments{
\item{m}{An nm object.}

\item{path}{Optional path to save the control file.}

\item{force}{Logical (default = \code{FALSE}), force write, don't ask.}
}
\value{
Invisibly returns \code{m} unmodified.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Not designed to be used directly.  \code{\link[=run_nm]{run_nm()}} and other functions will use
this to write the control file contents to the \code{ctl_path()} prior to
execution.
}
\keyword{internal}
