% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{write_ctl}
\alias{write_ctl}
\title{Write control file to disk}
\usage{
write_ctl(m, force = FALSE)
}
\arguments{
\item{m}{nm object}

\item{force}{logical (default = FALSE), force write, don't ask (ignore behaviour)}
}
\description{
Normally used by other functions
}
