% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctl_handling.R
\name{write_ctl}
\alias{write_ctl}
\title{write control file}
\usage{
write_ctl(ctl, dest, dir = getOption("models.dir"))
}
\arguments{
\item{ctl}{object of class character, ctl_character, or ctl_list}

\item{dest}{character or numeric. new run_id}

\item{dir}{character. Directory to place file. Default = getOption("models.dir")}
}
\description{
write control file
}
