% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMprojectExtra.R
\name{write_derived_data}
\alias{write_derived_data}
\title{Write derived data file.}
\usage{
write_derived_data(d, name, ...)
}
\arguments{
\item{d}{data.frame. Data frame to be saved}

\item{name}{name of file (without extension). If not a path, will save to
DerivedData directory}

\item{...}{additional arguments to be passed to write.csv}
}
\description{
will write to (DerivedData) directory
}
